+++ docs/assets/checkLogin-dd8Ug7jR.js
@@ -1,0 +1,1 @@
+import{g as a,a as n,p as o}from"./header-tMjJ4wC4.js";import{g as s,d as c,s as r}from"./storage-PSdSBPgg.js";function p(t,e){typeof t=="string"&&(t=a(t)),t.insertAdjacentHTML("beforebegin",e)}function h(t,e){typeof t=="string"&&(t=a(t)),t.insertAdjacentHTML("afterbegin",e)}function b(t,e){typeof t=="string"&&(t=a(t)),t.insertAdjacentHTML("beforeend",e)}function d(t){return t.toString().replace(/\B(?=(\d{3})+(?!\d))/g,",")}function k(t,e="image",i=0){return Array.isArray(t.fileName)?`https://square-zebra.pockethost.io/api/files/${t.collectionId}/${t.id}/${t[e][i]}`:`https://square-zebra.pockethost.io/api/files/${t.collectionId}/${t.id}/${t[e]}`}const f={isAuth:!1,user:null,token:""},l={user:null,id:[]};async function y(){if(await s("pocketbase_auth")){const{model:t}=await s("pocketbase_auth"),e=n(".header_member_service li a");e[0].textContent=`${t.name} 님`,e[0].style.color="var(--black)",e[0].style.fontWeight="700",e[1].textContent="로그아웃",e[1].href="/Karly/src/pages/main/",e[1].addEventListener("click",function(){o.authStore.clear(),c("pocketbase_auth"),r("view",l),r("auth",f)})}}export{f as a,d as b,y as c,l as d,p as e,h as f,k as g,b as i};
