--- docs/assets/getPbImageURL-Q86jnK4B.js
@@ -1,1 +1,0 @@
-import{a as e}from"./header-MICfn_ID.js";function a(t,r){typeof t=="string"&&(t=e(t)),t.insertAdjacentHTML("beforebegin",r)}function n(t,r){typeof t=="string"&&(t=e(t)),t.insertAdjacentHTML("afterbegin",r)}function f(t,r){typeof t=="string"&&(t=e(t)),t.insertAdjacentHTML("beforeend",r)}function o(t){return t.toString().replace(/\B(?=(\d{3})+(?!\d))/g,",")}function c(t,r="image",i=0){return Array.isArray(t.fileName)?`https://square-zebra.pockethost.io/api/files/${t.collectionId}/${t.id}/${t[r][i]}`:`https://square-zebra.pockethost.io/api/files/${t.collectionId}/${t.id}/${t[r]}`}export{a,n as b,o as c,c as g,f as i};
