--- docs/assets/header-MICfn_ID.js
@@ -1,1 +1,0 @@
-import"./style-v06b81ne.js";function r(t,e=document){if(typeof t!="string")throw new Error("getNode 함수의 인수는 문자 타입 이어야 합니다.");return e.nodeType!==document.DOCUMENT_NODE&&(e=document.querySelector(e)),e.querySelector(t)}function b(t){if(typeof t!="string")throw new Error("getNodes 함수의 인수는 문자 타입 이어야 합니다.");return document.querySelectorAll(t)}function g(t,e){if(typeof t=="string"&&(t=r(t)),typeof e!="string")throw new TypeError("addClass 함수의 두 번째 인수는 문자 타입 이어야 합니다.");t.classList.add(e)}function w(t,e){if(typeof t=="string"&&(t=r(t)),!e){t.className="";return}if(typeof e!="string")throw new TypeError("removeClass 함수의 두 번째 인수는 문자 타입 이어야 합니다.");t.classList.remove(e)}function m(t,e){if(typeof t=="string"&&(t=r(t)),!(e in document.body.style))throw new SyntaxError("getCss 함수의 두 번째 인수는 유효한 css 속성이 아닙니다.");return getComputedStyle(t).getPropertyValue(e)}function p(t,e,s){if(typeof t=="string"&&(t=r(t)),!(e in document.body.style))throw new SyntaxError("setCss 함수의 두 번째 인수는 유효한 css 속성이 아닙니다.");if(!s)throw new SyntaxError("setCss 함수의 세 번째 인수는 필수값입니다.");t.style[e]=s}const c=(t,e,s)=>s?p(t,e,s):m(t,e);function i(t){return Object.prototype.toString.call(t).slice(8,-1).toLowerCase()}const h=t=>i(t)==="object",C=t=>i(t)==="number",O=t=>i(t)==="string",l={method:"GET",body:null,headers:{"Content-Type":"application/json","Access-Control-Allow-Origin":"*"}},o=async t=>{const{url:e,...s}={...l,...t,headers:{...l.headers,...t.headers}},n=await fetch(e,s);return n.ok&&(n.data=await n.json()),n};o.get=(t,e)=>o({url:t,...e});o.post=(t,e,s)=>o({method:"POST",url:t,body:JSON.stringify(e),...s});o.delete=(t,e)=>o({method:"DELETE",url:t,...e});o.put=(t,e,s)=>o({method:"PUT",url:t,body:JSON.stringify(e),...s});r(".first");r(".second");const a={shouldReject:!1,timeout:1e3,data:"아싸 성공!",errorMessage:"알 수 없는 오류가 발생했습니다."};function E(t){let e={...a};C(t)&&(e.timeout=t),h(t)&&(e={...a,...t});let{timeout:s,shouldReject:n,errorMessage:f,data:u}=e;return new Promise((y,d)=>{setTimeout(()=>{n?d({message:f}):y(u)},s)})}E(1e3);(function(){const t=r(".nav_scroll");window.addEventListener("scroll",()=>{window.scrollY>=190?(c(t,"display","flex"),g(t,"sticky")):window.scrollY<190&&(c(t,"display","none"),w(t,"sticky"))});const e=r(".top_banner"),s=r(".top_banner_close");function n(){e.style.transition="All 0.5s",e.style.transform="translateY(-50px)",setTimeout(()=>{e.remove()},250)}s.addEventListener("click",n)})();export{r as a,c,b as g,O as i,o as t};
