+++ docs/assets/header-tMjJ4wC4.js
@@ -1,0 +1,1 @@
+import"./style-v06b81ne.js";class p extends Error{constructor(e){super("ClientResponseError"),this.url="",this.status=0,this.response={},this.isAbort=!1,this.originalError=null,Object.setPrototypeOf(this,p.prototype),e!==null&&typeof e=="object"&&(this.url=typeof e.url=="string"?e.url:"",this.status=typeof e.status=="number"?e.status:0,this.isAbort=!!e.isAbort,this.originalError=e.originalError,e.response!==null&&typeof e.response=="object"?this.response=e.response:e.data!==null&&typeof e.data=="object"?this.response=e.data:this.response={}),this.originalError||e instanceof p||(this.originalError=e),typeof DOMException<"u"&&e instanceof DOMException&&(this.isAbort=!0),this.name="ClientResponseError "+this.status,this.message=this.response?.message,this.message||(this.isAbort?this.message="The request was autocancelled. You can find more info in https://github.com/pocketbase/js-sdk#auto-cancellation.":this.originalError?.cause?.message?.includes("ECONNREFUSED ::1")?this.message="Failed to connect to the PocketBase server. Try changing the SDK URL from localhost to 127.0.0.1 (https://github.com/pocketbase/js-sdk/issues/21).":this.message="Something went wrong while processing your request.")}get data(){return this.response}toJSON(){return{...this}}}const T=/^[\u0009\u0020-\u007e\u0080-\u00ff]+$/;function W(n,e){const t={};if(typeof n!="string")return t;const s=Object.assign({},e||{}).decode||$;let i=0;for(;i<n.length;){const r=n.indexOf("=",i);if(r===-1)break;let o=n.indexOf(";",i);if(o===-1)o=n.length;else if(o<r){i=n.lastIndexOf(";",r-1)+1;continue}const a=n.slice(i,r).trim();if(t[a]===void 0){let c=n.slice(r+1,o).trim();c.charCodeAt(0)===34&&(c=c.slice(1,-1));try{t[a]=s(c)}catch{t[a]=c}}i=o+1}return t}function j(n,e,t){const s=Object.assign({},t||{}),i=s.encode||M;if(!T.test(n))throw new TypeError("argument name is invalid");const r=i(e);if(r&&!T.test(r))throw new TypeError("argument val is invalid");let o=n+"="+r;if(s.maxAge!=null){const a=s.maxAge-0;if(isNaN(a)||!isFinite(a))throw new TypeError("option maxAge is invalid");o+="; Max-Age="+Math.floor(a)}if(s.domain){if(!T.test(s.domain))throw new TypeError("option domain is invalid");o+="; Domain="+s.domain}if(s.path){if(!T.test(s.path))throw new TypeError("option path is invalid");o+="; Path="+s.path}if(s.expires){if(!function(c){return Object.prototype.toString.call(c)==="[object Date]"||c instanceof Date}(s.expires)||isNaN(s.expires.valueOf()))throw new TypeError("option expires is invalid");o+="; Expires="+s.expires.toUTCString()}if(s.httpOnly&&(o+="; HttpOnly"),s.secure&&(o+="; Secure"),s.priority)switch(typeof s.priority=="string"?s.priority.toLowerCase():s.priority){case"low":o+="; Priority=Low";break;case"medium":o+="; Priority=Medium";break;case"high":o+="; Priority=High";break;default:throw new TypeError("option priority is invalid")}if(s.sameSite)switch(typeof s.sameSite=="string"?s.sameSite.toLowerCase():s.sameSite){case!0:o+="; SameSite=Strict";break;case"lax":o+="; SameSite=Lax";break;case"strict":o+="; SameSite=Strict";break;case"none":o+="; SameSite=None";break;default:throw new TypeError("option sameSite is invalid")}return o}function $(n){return n.indexOf("%")!==-1?decodeURIComponent(n):n}function M(n){return encodeURIComponent(n)}let D;function v(n){if(n)try{const e=decodeURIComponent(D(n.split(".")[1]).split("").map(function(t){return"%"+("00"+t.charCodeAt(0).toString(16)).slice(-2)}).join(""));return JSON.parse(e)||{}}catch{}return{}}function _(n,e=0){let t=v(n);return!(Object.keys(t).length>0&&(!t.exp||t.exp-e>Date.now()/1e3))}D=typeof atob=="function"?atob:n=>{let e=String(n).replace(/=+$/,"");if(e.length%4==1)throw new Error("'atob' failed: The string to be decoded is not correctly encoded.");for(var t,s,i=0,r=0,o="";s=e.charAt(r++);~s&&(t=i%4?64*t+s:s,i++%4)?o+=String.fromCharCode(255&t>>(-2*i&6)):0)s="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=".indexOf(s);return o};const q="pb_auth";class G{constructor(){this.baseToken="",this.baseModel=null,this._onChangeCallbacks=[]}get token(){return this.baseToken}get model(){return this.baseModel}get isValid(){return!_(this.token)}get isAdmin(){return v(this.token).type==="admin"}get isAuthRecord(){return v(this.token).type==="authRecord"}save(e,t){this.baseToken=e||"",this.baseModel=t||null,this.triggerChange()}clear(){this.baseToken="",this.baseModel=null,this.triggerChange()}loadFromCookie(e,t=q){const s=W(e||"")[t]||"";let i={};try{i=JSON.parse(s),(typeof i===null||typeof i!="object"||Array.isArray(i))&&(i={})}catch{}this.save(i.token||"",i.model||null)}exportToCookie(e,t=q){const s={secure:!0,sameSite:!0,httpOnly:!0,path:"/"},i=v(this.token);s.expires=i?.exp?new Date(1e3*i.exp):new Date("1970-01-01"),e=Object.assign({},s,e);const r={token:this.token,model:this.model?JSON.parse(JSON.stringify(this.model)):null};let o=j(t,JSON.stringify(r),e);const a=typeof Blob<"u"?new Blob([o]).size:o.length;if(r.model&&a>4096){r.model={id:r?.model?.id,email:r?.model?.email};const c=["collectionId","username","verified"];for(const l in this.model)c.includes(l)&&(r.model[l]=this.model[l]);o=j(t,JSON.stringify(r),e)}return o}onChange(e,t=!1){return this._onChangeCallbacks.push(e),t&&e(this.token,this.model),()=>{for(let s=this._onChangeCallbacks.length-1;s>=0;s--)if(this._onChangeCallbacks[s]==e)return delete this._onChangeCallbacks[s],void this._onChangeCallbacks.splice(s,1)}}triggerChange(){for(const e of this._onChangeCallbacks)e&&e(this.token,this.model)}}class J extends G{constructor(e="pocketbase_auth"){super(),this.storageFallback={},this.storageKey=e,this._bindStorageEvent()}get token(){return(this._storageGet(this.storageKey)||{}).token||""}get model(){return(this._storageGet(this.storageKey)||{}).model||null}save(e,t){this._storageSet(this.storageKey,{token:e,model:t}),super.save(e,t)}clear(){this._storageRemove(this.storageKey),super.clear()}_storageGet(e){if(typeof window<"u"&&window?.localStorage){const t=window.localStorage.getItem(e)||"";try{return JSON.parse(t)}catch{return t}}return this.storageFallback[e]}_storageSet(e,t){if(typeof window<"u"&&window?.localStorage){let s=t;typeof t!="string"&&(s=JSON.stringify(t)),window.localStorage.setItem(e,s)}else this.storageFallback[e]=t}_storageRemove(e){typeof window<"u"&&window?.localStorage&&window.localStorage?.removeItem(e),delete this.storageFallback[e]}_bindStorageEvent(){typeof window<"u"&&window?.localStorage&&window.addEventListener&&window.addEventListener("storage",e=>{if(e.key!=this.storageKey)return;const t=this._storageGet(this.storageKey)||{};super.save(t.token||"",t.model||null)})}}class y{constructor(e){this.client=e}}class H extends y{getAll(e){return e=Object.assign({method:"GET"},e),this.client.send("/api/settings",e)}update(e,t){return t=Object.assign({method:"PATCH",body:e},t),this.client.send("/api/settings",t)}testS3(e="storage",t){return t=Object.assign({method:"POST",body:{filesystem:e}},t),this.client.send("/api/settings/test/s3",t).then(()=>!0)}testEmail(e,t,s){return s=Object.assign({method:"POST",body:{email:e,template:t}},s),this.client.send("/api/settings/test/email",s).then(()=>!0)}generateAppleClientSecret(e,t,s,i,r,o){return o=Object.assign({method:"POST",body:{clientId:e,teamId:t,keyId:s,privateKey:i,duration:r}},o),this.client.send("/api/settings/apple/generate-client-secret",o)}}class I extends y{decode(e){return e}getFullList(e,t){if(typeof e=="number")return this._getFullList(e,t);let s=500;return(t=Object.assign({},e,t)).batch&&(s=t.batch,delete t.batch),this._getFullList(s,t)}getList(e=1,t=30,s){return(s=Object.assign({method:"GET"},s)).query=Object.assign({page:e,perPage:t},s.query),this.client.send(this.baseCrudPath,s).then(i=>(i.items=i.items?.map(r=>this.decode(r))||[],i))}getFirstListItem(e,t){return(t=Object.assign({requestKey:"one_by_filter_"+this.baseCrudPath+"_"+e},t)).query=Object.assign({filter:e,skipTotal:1},t.query),this.getList(1,1,t).then(s=>{if(!s?.items?.length)throw new p({status:404,data:{code:404,message:"The requested resource wasn't found.",data:{}}});return s.items[0]})}getOne(e,t){return t=Object.assign({method:"GET"},t),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(e),t).then(s=>this.decode(s))}create(e,t){return t=Object.assign({method:"POST",body:e},t),this.client.send(this.baseCrudPath,t).then(s=>this.decode(s))}update(e,t,s){return s=Object.assign({method:"PATCH",body:t},s),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(e),s).then(i=>this.decode(i))}delete(e,t){return t=Object.assign({method:"DELETE"},t),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(e),t).then(()=>!0)}_getFullList(e=500,t){(t=t||{}).query=Object.assign({skipTotal:1},t.query);let s=[],i=async r=>this.getList(r,e||500,t).then(o=>{const a=o.items;return s=s.concat(a),a.length==o.perPage?i(r+1):s});return i(1)}}function h(n,e,t,s){const i=s!==void 0;return i||t!==void 0?i?(console.warn(n),e.body=Object.assign({},e.body,t),e.query=Object.assign({},e.query,s),e):Object.assign(e,t):e}function E(n){n._resetAutoRefresh?.()}class V extends I{get baseCrudPath(){return"/api/admins"}update(e,t,s){return super.update(e,t,s).then(i=>(this.client.authStore.model?.id===i.id&&this.client.authStore.model?.collectionId===void 0&&this.client.authStore.save(this.client.authStore.token,i),i))}delete(e,t){return super.delete(e,t).then(s=>(s&&this.client.authStore.model?.id===e&&this.client.authStore.model?.collectionId===void 0&&this.client.authStore.clear(),s))}authResponse(e){const t=this.decode(e?.admin||{});return e?.token&&e?.admin&&this.client.authStore.save(e.token,t),Object.assign({},e,{token:e?.token||"",admin:t})}async authWithPassword(e,t,s,i){let r={method:"POST",body:{identity:e,password:t}};r=h("This form of authWithPassword(email, pass, body?, query?) is deprecated. Consider replacing it with authWithPassword(email, pass, options?).",r,s,i);const o=r.autoRefreshThreshold;delete r.autoRefreshThreshold,r.autoRefresh||E(this.client);let a=await this.client.send(this.baseCrudPath+"/auth-with-password",r);return a=this.authResponse(a),o&&function(l,w,k,A){E(l);const d=l.beforeSend,S=l.authStore.model,f=l.authStore.onChange((b,u)=>{(!b||u?.id!=S?.id||(u?.collectionId||S?.collectionId)&&u?.collectionId!=S?.collectionId)&&E(l)});l._resetAutoRefresh=function(){f(),l.beforeSend=d,delete l._resetAutoRefresh},l.beforeSend=async(b,u)=>{const B=l.authStore.token;if(u.query?.autoRefresh)return d?d(b,u):{url:b,sendOptions:u};let P=l.authStore.isValid;if(P&&_(l.authStore.token,w))try{await k()}catch{P=!1}P||await A();const C=u.headers||{};for(let O in C)if(O.toLowerCase()=="authorization"&&B==C[O]&&l.authStore.token){C[O]=l.authStore.token;break}return u.headers=C,d?d(b,u):{url:b,sendOptions:u}}}(this.client,o,()=>this.authRefresh({autoRefresh:!0}),()=>this.authWithPassword(e,t,Object.assign({autoRefresh:!0},r))),a}authRefresh(e,t){let s={method:"POST"};return s=h("This form of authRefresh(body?, query?) is deprecated. Consider replacing it with authRefresh(options?).",s,e,t),this.client.send(this.baseCrudPath+"/auth-refresh",s).then(this.authResponse.bind(this))}requestPasswordReset(e,t,s){let i={method:"POST",body:{email:e}};return i=h("This form of requestPasswordReset(email, body?, query?) is deprecated. Consider replacing it with requestPasswordReset(email, options?).",i,t,s),this.client.send(this.baseCrudPath+"/request-password-reset",i).then(()=>!0)}confirmPasswordReset(e,t,s,i,r){let o={method:"POST",body:{token:e,password:t,passwordConfirm:s}};return o=h("This form of confirmPasswordReset(resetToken, password, passwordConfirm, body?, query?) is deprecated. Consider replacing it with confirmPasswordReset(resetToken, password, passwordConfirm, options?).",o,i,r),this.client.send(this.baseCrudPath+"/confirm-password-reset",o).then(()=>!0)}}const z=["requestKey","$cancelKey","$autoCancel","fetch","headers","body","query","params","cache","credentials","headers","integrity","keepalive","method","mode","redirect","referrer","referrerPolicy","signal","window"];function F(n){if(n){n.query=n.query||{};for(let e in n)z.includes(e)||(n.query[e]=n[e],delete n[e])}}class K extends y{constructor(){super(...arguments),this.clientId="",this.eventSource=null,this.subscriptions={},this.lastSentSubscriptions=[],this.maxConnectTimeout=15e3,this.reconnectAttempts=0,this.maxReconnectAttempts=1/0,this.predefinedReconnectIntervals=[200,300,500,1e3,1200,1500,2e3],this.pendingConnects=[]}get isConnected(){return!!this.eventSource&&!!this.clientId&&!this.pendingConnects.length}async subscribe(e,t,s){if(!e)throw new Error("topic must be set.");let i=e;if(s){F(s);const o="options="+encodeURIComponent(JSON.stringify({query:s.query,headers:s.headers}));i+=(i.includes("?")?"&":"?")+o}const r=function(o){const a=o;let c;try{c=JSON.parse(a?.data)}catch{}t(c||{})};return this.subscriptions[i]||(this.subscriptions[i]=[]),this.subscriptions[i].push(r),this.isConnected?this.subscriptions[i].length===1?await this.submitSubscriptions():this.eventSource?.addEventListener(i,r):await this.connect(),async()=>this.unsubscribeByTopicAndListener(e,r)}async unsubscribe(e){let t=!1;if(e){const s=this.getSubscriptionsByTopic(e);for(let i in s)if(this.hasSubscriptionListeners(i)){for(let r of this.subscriptions[i])this.eventSource?.removeEventListener(i,r);delete this.subscriptions[i],t||(t=!0)}}else this.subscriptions={};this.hasSubscriptionListeners()?t&&await this.submitSubscriptions():this.disconnect()}async unsubscribeByPrefix(e){let t=!1;for(let s in this.subscriptions)if((s+"?").startsWith(e)){t=!0;for(let i of this.subscriptions[s])this.eventSource?.removeEventListener(s,i);delete this.subscriptions[s]}t&&(this.hasSubscriptionListeners()?await this.submitSubscriptions():this.disconnect())}async unsubscribeByTopicAndListener(e,t){let s=!1;const i=this.getSubscriptionsByTopic(e);for(let r in i){if(!Array.isArray(this.subscriptions[r])||!this.subscriptions[r].length)continue;let o=!1;for(let a=this.subscriptions[r].length-1;a>=0;a--)this.subscriptions[r][a]===t&&(o=!0,delete this.subscriptions[r][a],this.subscriptions[r].splice(a,1),this.eventSource?.removeEventListener(r,t));o&&(this.subscriptions[r].length||delete this.subscriptions[r],s||this.hasSubscriptionListeners(r)||(s=!0))}this.hasSubscriptionListeners()?s&&await this.submitSubscriptions():this.disconnect()}hasSubscriptionListeners(e){if(this.subscriptions=this.subscriptions||{},e)return!!this.subscriptions[e]?.length;for(let t in this.subscriptions)if(this.subscriptions[t]?.length)return!0;return!1}async submitSubscriptions(){if(this.clientId)return this.addAllSubscriptionListeners(),this.lastSentSubscriptions=this.getNonEmptySubscriptionKeys(),this.client.send("/api/realtime",{method:"POST",body:{clientId:this.clientId,subscriptions:this.lastSentSubscriptions},requestKey:this.getSubscriptionsCancelKey()}).catch(e=>{if(!e?.isAbort)throw e})}getSubscriptionsCancelKey(){return"realtime_"+this.clientId}getSubscriptionsByTopic(e){const t={};e=e.includes("?")?e:e+"?";for(let s in this.subscriptions)(s+"?").startsWith(e)&&(t[s]=this.subscriptions[s]);return t}getNonEmptySubscriptionKeys(){const e=[];for(let t in this.subscriptions)this.subscriptions[t].length&&e.push(t);return e}addAllSubscriptionListeners(){if(this.eventSource){this.removeAllSubscriptionListeners();for(let e in this.subscriptions)for(let t of this.subscriptions[e])this.eventSource.addEventListener(e,t)}}removeAllSubscriptionListeners(){if(this.eventSource)for(let e in this.subscriptions)for(let t of this.subscriptions[e])this.eventSource.removeEventListener(e,t)}async connect(){if(!(this.reconnectAttempts>0))return new Promise((e,t)=>{this.pendingConnects.push({resolve:e,reject:t}),this.pendingConnects.length>1||this.initConnect()})}initConnect(){this.disconnect(!0),clearTimeout(this.connectTimeoutId),this.connectTimeoutId=setTimeout(()=>{this.connectErrorHandler(new Error("EventSource connect took too long."))},this.maxConnectTimeout),this.eventSource=new EventSource(this.client.buildUrl("/api/realtime")),this.eventSource.onerror=e=>{this.connectErrorHandler(new Error("Failed to establish realtime connection."))},this.eventSource.addEventListener("PB_CONNECT",e=>{const t=e;this.clientId=t?.lastEventId,this.submitSubscriptions().then(async()=>{let s=3;for(;this.hasUnsentSubscriptions()&&s>0;)s--,await this.submitSubscriptions()}).then(()=>{for(let i of this.pendingConnects)i.resolve();this.pendingConnects=[],this.reconnectAttempts=0,clearTimeout(this.reconnectTimeoutId),clearTimeout(this.connectTimeoutId);const s=this.getSubscriptionsByTopic("PB_CONNECT");for(let i in s)for(let r of s[i])r(e)}).catch(s=>{this.clientId="",this.connectErrorHandler(s)})})}hasUnsentSubscriptions(){const e=this.getNonEmptySubscriptionKeys();if(e.length!=this.lastSentSubscriptions.length)return!0;for(const t of e)if(!this.lastSentSubscriptions.includes(t))return!0;return!1}connectErrorHandler(e){if(clearTimeout(this.connectTimeoutId),clearTimeout(this.reconnectTimeoutId),!this.clientId&&!this.reconnectAttempts||this.reconnectAttempts>this.maxReconnectAttempts){for(let s of this.pendingConnects)s.reject(new p(e));return this.pendingConnects=[],void this.disconnect()}this.disconnect(!0);const t=this.predefinedReconnectIntervals[this.reconnectAttempts]||this.predefinedReconnectIntervals[this.predefinedReconnectIntervals.length-1];this.reconnectAttempts++,this.reconnectTimeoutId=setTimeout(()=>{this.initConnect()},t)}disconnect(e=!1){if(clearTimeout(this.connectTimeoutId),clearTimeout(this.reconnectTimeoutId),this.removeAllSubscriptionListeners(),this.client.cancelRequest(this.getSubscriptionsCancelKey()),this.eventSource?.close(),this.eventSource=null,this.clientId="",!e){this.reconnectAttempts=0;for(let t of this.pendingConnects)t.resolve();this.pendingConnects=[]}}}class Q extends I{constructor(e,t){super(e),this.collectionIdOrName=t}get baseCrudPath(){return this.baseCollectionPath+"/records"}get baseCollectionPath(){return"/api/collections/"+encodeURIComponent(this.collectionIdOrName)}async subscribe(e,t,s){if(!e)throw new Error("Missing topic.");if(!t)throw new Error("Missing subscription callback.");return this.client.realtime.subscribe(this.collectionIdOrName+"/"+e,t,s)}async unsubscribe(e){return e?this.client.realtime.unsubscribe(this.collectionIdOrName+"/"+e):this.client.realtime.unsubscribeByPrefix(this.collectionIdOrName)}getFullList(e,t){if(typeof e=="number")return super.getFullList(e,t);const s=Object.assign({},e,t);return super.getFullList(s)}getList(e=1,t=30,s){return super.getList(e,t,s)}getFirstListItem(e,t){return super.getFirstListItem(e,t)}getOne(e,t){return super.getOne(e,t)}create(e,t){return super.create(e,t)}update(e,t,s){return super.update(e,t,s).then(i=>(this.client.authStore.model?.id!==i?.id||this.client.authStore.model?.collectionId!==this.collectionIdOrName&&this.client.authStore.model?.collectionName!==this.collectionIdOrName||this.client.authStore.save(this.client.authStore.token,i),i))}delete(e,t){return super.delete(e,t).then(s=>(!s||this.client.authStore.model?.id!==e||this.client.authStore.model?.collectionId!==this.collectionIdOrName&&this.client.authStore.model?.collectionName!==this.collectionIdOrName||this.client.authStore.clear(),s))}authResponse(e){const t=this.decode(e?.record||{});return this.client.authStore.save(e?.token,t),Object.assign({},e,{token:e?.token||"",record:t})}listAuthMethods(e){return e=Object.assign({method:"GET"},e),this.client.send(this.baseCollectionPath+"/auth-methods",e).then(t=>Object.assign({},t,{usernamePassword:!!t?.usernamePassword,emailPassword:!!t?.emailPassword,authProviders:Array.isArray(t?.authProviders)?t?.authProviders:[]}))}authWithPassword(e,t,s,i){let r={method:"POST",body:{identity:e,password:t}};return r=h("This form of authWithPassword(usernameOrEmail, pass, body?, query?) is deprecated. Consider replacing it with authWithPassword(usernameOrEmail, pass, options?).",r,s,i),this.client.send(this.baseCollectionPath+"/auth-with-password",r).then(o=>this.authResponse(o))}authWithOAuth2Code(e,t,s,i,r,o,a){let c={method:"POST",body:{provider:e,code:t,codeVerifier:s,redirectUrl:i,createData:r}};return c=h("This form of authWithOAuth2Code(provider, code, codeVerifier, redirectUrl, createData?, body?, query?) is deprecated. Consider replacing it with authWithOAuth2Code(provider, code, codeVerifier, redirectUrl, createData?, options?).",c,o,a),this.client.send(this.baseCollectionPath+"/auth-with-oauth2",c).then(l=>this.authResponse(l))}async authWithOAuth2(...e){if(e.length>1||typeof e?.[0]=="string")return console.warn("PocketBase: This form of authWithOAuth2() is deprecated and may get removed in the future. Please replace with authWithOAuth2Code() OR use the authWithOAuth2() realtime form as shown in https://pocketbase.io/docs/authentication/#oauth2-integration."),this.authWithOAuth2Code(e?.[0]||"",e?.[1]||"",e?.[2]||"",e?.[3]||"",e?.[4]||{},e?.[5]||{},e?.[6]||{});const t=e?.[0]||{},s=(await this.listAuthMethods()).authProviders.find(c=>c.name===t.provider);if(!s)throw new p(new Error(`Missing or invalid provider "${t.provider}".`));const i=this.client.buildUrl("/api/oauth2-redirect"),r=new K(this.client);let o=null;function a(){o?.close(),r.unsubscribe()}return t.urlCallback||(o=L(void 0)),new Promise(async(c,l)=>{try{await r.subscribe("@oauth2",async d=>{const S=r.clientId;try{if(!d.state||S!==d.state)throw new Error("State parameters don't match.");const f=Object.assign({},t);delete f.provider,delete f.scopes,delete f.createData,delete f.urlCallback;const b=await this.authWithOAuth2Code(s.name,d.code,s.codeVerifier,i,t.createData,f);c(b)}catch(f){l(new p(f))}a()});const w={state:r.clientId};t.scopes?.length&&(w.scope=t.scopes.join(" "));const k=this._replaceQueryParams(s.authUrl+i,w);await(t.urlCallback||function(d){o?o.location.href=d:o=L(d)})(k)}catch(w){a(),l(new p(w))}})}authRefresh(e,t){let s={method:"POST"};return s=h("This form of authRefresh(body?, query?) is deprecated. Consider replacing it with authRefresh(options?).",s,e,t),this.client.send(this.baseCollectionPath+"/auth-refresh",s).then(i=>this.authResponse(i))}requestPasswordReset(e,t,s){let i={method:"POST",body:{email:e}};return i=h("This form of requestPasswordReset(email, body?, query?) is deprecated. Consider replacing it with requestPasswordReset(email, options?).",i,t,s),this.client.send(this.baseCollectionPath+"/request-password-reset",i).then(()=>!0)}confirmPasswordReset(e,t,s,i,r){let o={method:"POST",body:{token:e,password:t,passwordConfirm:s}};return o=h("This form of confirmPasswordReset(token, password, passwordConfirm, body?, query?) is deprecated. Consider replacing it with confirmPasswordReset(token, password, passwordConfirm, options?).",o,i,r),this.client.send(this.baseCollectionPath+"/confirm-password-reset",o).then(()=>!0)}requestVerification(e,t,s){let i={method:"POST",body:{email:e}};return i=h("This form of requestVerification(email, body?, query?) is deprecated. Consider replacing it with requestVerification(email, options?).",i,t,s),this.client.send(this.baseCollectionPath+"/request-verification",i).then(()=>!0)}confirmVerification(e,t,s){let i={method:"POST",body:{token:e}};return i=h("This form of confirmVerification(token, body?, query?) is deprecated. Consider replacing it with confirmVerification(token, options?).",i,t,s),this.client.send(this.baseCollectionPath+"/confirm-verification",i).then(()=>!0)}requestEmailChange(e,t,s){let i={method:"POST",body:{newEmail:e}};return i=h("This form of requestEmailChange(newEmail, body?, query?) is deprecated. Consider replacing it with requestEmailChange(newEmail, options?).",i,t,s),this.client.send(this.baseCollectionPath+"/request-email-change",i).then(()=>!0)}confirmEmailChange(e,t,s,i){let r={method:"POST",body:{token:e,password:t}};return r=h("This form of confirmEmailChange(token, password, body?, query?) is deprecated. Consider replacing it with confirmEmailChange(token, password, options?).",r,s,i),this.client.send(this.baseCollectionPath+"/confirm-email-change",r).then(()=>!0)}listExternalAuths(e,t){return t=Object.assign({method:"GET"},t),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(e)+"/external-auths",t)}unlinkExternalAuth(e,t,s){return s=Object.assign({method:"DELETE"},s),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(e)+"/external-auths/"+encodeURIComponent(t),s).then(()=>!0)}_replaceQueryParams(e,t={}){let s=e,i="";e.indexOf("?")>=0&&(s=e.substring(0,e.indexOf("?")),i=e.substring(e.indexOf("?")+1));const r={},o=i.split("&");for(const a of o){if(a=="")continue;const c=a.split("=");r[decodeURIComponent(c[0].replace(/\+/g," "))]=decodeURIComponent((c[1]||"").replace(/\+/g," "))}for(let a in t)t.hasOwnProperty(a)&&(t[a]==null?delete r[a]:r[a]=t[a]);i="";for(let a in r)r.hasOwnProperty(a)&&(i!=""&&(i+="&"),i+=encodeURIComponent(a.replace(/%20/g,"+"))+"="+encodeURIComponent(r[a].replace(/%20/g,"+")));return i!=""?s+"?"+i:s}}function L(n){if(typeof window>"u"||!window?.open)throw new p(new Error("Not in a browser context - please pass a custom urlCallback function."));let e=1024,t=768,s=window.innerWidth,i=window.innerHeight;e=e>s?s:e,t=t>i?i:t;let r=s/2-e/2,o=i/2-t/2;return window.open(n,"popup_window","width="+e+",height="+t+",top="+o+",left="+r+",resizable,menubar=no")}class Y extends I{get baseCrudPath(){return"/api/collections"}async import(e,t=!1,s){return s=Object.assign({method:"PUT",body:{collections:e,deleteMissing:t}},s),this.client.send(this.baseCrudPath+"/import",s).then(()=>!0)}}class X extends y{getList(e=1,t=30,s){return(s=Object.assign({method:"GET"},s)).query=Object.assign({page:e,perPage:t},s.query),this.client.send("/api/logs",s)}getOne(e,t){return t=Object.assign({method:"GET"},t),this.client.send("/api/logs/"+encodeURIComponent(e),t)}getStats(e){return e=Object.assign({method:"GET"},e),this.client.send("/api/logs/stats",e)}}class Z extends y{check(e){return e=Object.assign({method:"GET"},e),this.client.send("/api/health",e)}}class ee extends y{getUrl(e,t,s={}){if(!t||!e?.id||!e?.collectionId&&!e?.collectionName)return"";const i=[];i.push("api"),i.push("files"),i.push(encodeURIComponent(e.collectionId||e.collectionName)),i.push(encodeURIComponent(e.id)),i.push(encodeURIComponent(t));let r=this.client.buildUrl(i.join("/"));if(Object.keys(s).length){s.download===!1&&delete s.download;const o=new URLSearchParams(s);r+=(r.includes("?")?"&":"?")+o}return r}getToken(e){return e=Object.assign({method:"POST"},e),this.client.send("/api/files/token",e).then(t=>t?.token||"")}}class te extends y{getFullList(e){return e=Object.assign({method:"GET"},e),this.client.send("/api/backups",e)}create(e,t){return t=Object.assign({method:"POST",body:{name:e}},t),this.client.send("/api/backups",t).then(()=>!0)}upload(e,t){return t=Object.assign({method:"POST",body:e},t),this.client.send("/api/backups/upload",t).then(()=>!0)}delete(e,t){return t=Object.assign({method:"DELETE"},t),this.client.send(`/api/backups/${encodeURIComponent(e)}`,t).then(()=>!0)}restore(e,t){return t=Object.assign({method:"POST"},t),this.client.send(`/api/backups/${encodeURIComponent(e)}/restore`,t).then(()=>!0)}getDownloadUrl(e,t){return this.client.buildUrl(`/api/backups/${encodeURIComponent(t)}?token=${encodeURIComponent(e)}`)}}class se{constructor(e="/",t,s="en-US"){this.cancelControllers={},this.recordServices={},this.enableAutoCancellation=!0,this.baseUrl=e,this.lang=s,this.authStore=t||new J,this.admins=new V(this),this.collections=new Y(this),this.files=new ee(this),this.logs=new X(this),this.settings=new H(this),this.realtime=new K(this),this.health=new Z(this),this.backups=new te(this)}collection(e){return this.recordServices[e]||(this.recordServices[e]=new Q(this,e)),this.recordServices[e]}autoCancellation(e){return this.enableAutoCancellation=!!e,this}cancelRequest(e){return this.cancelControllers[e]&&(this.cancelControllers[e].abort(),delete this.cancelControllers[e]),this}cancelAllRequests(){for(let e in this.cancelControllers)this.cancelControllers[e].abort();return this.cancelControllers={},this}filter(e,t){if(!t)return e;for(let s in t){let i=t[s];switch(typeof i){case"boolean":case"number":i=""+i;break;case"string":i="'"+i.replace(/'/g,"\\'")+"'";break;default:i=i===null?"null":i instanceof Date?"'"+i.toISOString().replace("T"," ")+"'":"'"+JSON.stringify(i).replace(/'/g,"\\'")+"'"}e=e.replaceAll("{:"+s+"}",i)}return e}getFileUrl(e,t,s={}){return this.files.getUrl(e,t,s)}buildUrl(e){let t=this.baseUrl;return typeof window>"u"||!window.location||t.startsWith("https://")||t.startsWith("http://")||(t=window.location.origin?.endsWith("/")?window.location.origin.substring(0,window.location.origin.length-1):window.location.origin||"",this.baseUrl.startsWith("/")||(t+=window.location.pathname||"/",t+=t.endsWith("/")?"":"/"),t+=this.baseUrl),e&&(t+=t.endsWith("/")?"":"/",t+=e.startsWith("/")?e.substring(1):e),t}async send(e,t){t=this.initSendOptions(e,t);let s=this.buildUrl(e);if(this.beforeSend){const i=Object.assign({},await this.beforeSend(s,t));i.url!==void 0||i.options!==void 0?(s=i.url||s,t=i.options||t):Object.keys(i).length&&(t=i,console?.warn&&console.warn("Deprecated format of beforeSend return: please use `return { url, options }`, instead of `return options`."))}if(t.query!==void 0){const i=this.serializeQueryParams(t.query);i&&(s+=(s.includes("?")?"&":"?")+i),delete t.query}return this.getHeader(t.headers,"Content-Type")=="application/json"&&t.body&&typeof t.body!="string"&&(t.body=JSON.stringify(t.body)),(t.fetch||fetch)(s,t).then(async i=>{let r={};try{r=await i.json()}catch{}if(this.afterSend&&(r=await this.afterSend(i,r)),i.status>=400)throw new p({url:i.url,status:i.status,data:r});return r}).catch(i=>{throw new p(i)})}initSendOptions(e,t){if((t=Object.assign({method:"GET"},t)).body=this.convertToFormDataIfNeeded(t.body),F(t),t.query=Object.assign({},t.params,t.query),t.requestKey===void 0&&(t.$autoCancel===!1||t.query.$autoCancel===!1?t.requestKey=null:(t.$cancelKey||t.query.$cancelKey)&&(t.requestKey=t.$cancelKey||t.query.$cancelKey)),delete t.$autoCancel,delete t.query.$autoCancel,delete t.$cancelKey,delete t.query.$cancelKey,this.getHeader(t.headers,"Content-Type")!==null||this.isFormData(t.body)||(t.headers=Object.assign({},t.headers,{"Content-Type":"application/json"})),this.getHeader(t.headers,"Accept-Language")===null&&(t.headers=Object.assign({},t.headers,{"Accept-Language":this.lang})),this.authStore.token&&this.getHeader(t.headers,"Authorization")===null&&(t.headers=Object.assign({},t.headers,{Authorization:this.authStore.token})),this.enableAutoCancellation&&t.requestKey!==null){const s=t.requestKey||(t.method||"GET")+e;delete t.requestKey,this.cancelRequest(s);const i=new AbortController;this.cancelControllers[s]=i,t.signal=i.signal}return t}convertToFormDataIfNeeded(e){if(typeof FormData>"u"||e===void 0||typeof e!="object"||e===null||this.isFormData(e)||!this.hasBlobField(e))return e;const t=new FormData;for(let s in e){const i=Array.isArray(e[s])?e[s]:[e[s]];for(let r of i)t.append(s,r)}return t}hasBlobField(e){for(let t in e){const s=Array.isArray(e[t])?e[t]:[e[t]];for(let i of s)if(typeof Blob<"u"&&i instanceof Blob||typeof File<"u"&&i instanceof File)return!0}return!1}getHeader(e,t){e=e||{},t=t.toLowerCase();for(let s in e)if(s.toLowerCase()==t)return e[s];return null}isFormData(e){return e&&(e.constructor.name==="FormData"||typeof FormData<"u"&&e instanceof FormData)}serializeQueryParams(e){const t=[];for(const s in e){if(e[s]===null)continue;const i=e[s],r=encodeURIComponent(s);if(Array.isArray(i))for(const o of i)t.push(r+"="+encodeURIComponent(o));else i instanceof Date?t.push(r+"="+encodeURIComponent(i.toISOString())):typeof i!==null&&typeof i=="object"?t.push(r+"="+encodeURIComponent(JSON.stringify(i))):t.push(r+"="+encodeURIComponent(i))}return t.join("&")}}const de=new se("https://square-zebra.pockethost.io");function m(n,e=document){if(typeof n!="string")throw new Error("getNode 함수의 인수는 문자 타입 이어야 합니다.");return e.nodeType!==document.DOCUMENT_NODE&&(e=document.querySelector(e)),e.querySelector(n)}function ue(n){if(typeof n!="string")throw new Error("getNodes 함수의 인수는 문자 타입 이어야 합니다.");return document.querySelectorAll(n)}function ie(n,e){if(typeof n=="string"&&(n=m(n)),typeof e!="string")throw new TypeError("addClass 함수의 두 번째 인수는 문자 타입 이어야 합니다.");n.classList.add(e)}function ne(n,e){if(typeof n=="string"&&(n=m(n)),!e){n.className="";return}if(typeof e!="string")throw new TypeError("removeClass 함수의 두 번째 인수는 문자 타입 이어야 합니다.");n.classList.remove(e)}function re(n,e){if(typeof n=="string"&&(n=m(n)),!(e in document.body.style))throw new SyntaxError("getCss 함수의 두 번째 인수는 유효한 css 속성이 아닙니다.");return getComputedStyle(n).getPropertyValue(e)}function oe(n,e,t){if(typeof n=="string"&&(n=m(n)),!(e in document.body.style))throw new SyntaxError("setCss 함수의 두 번째 인수는 유효한 css 속성이 아닙니다.");if(!t)throw new SyntaxError("setCss 함수의 세 번째 인수는 필수값입니다.");n.style[e]=t}const U=(n,e,t)=>t?oe(n,e,t):re(n,e);function R(n){return Object.prototype.toString.call(n).slice(8,-1).toLowerCase()}const ae=n=>R(n)==="object",ce=n=>R(n)==="number",pe=n=>R(n)==="string",x={method:"GET",body:null,headers:{"Content-Type":"application/json","Access-Control-Allow-Origin":"*"}},g=async n=>{const{url:e,...t}={...x,...n,headers:{...x.headers,...n.headers}},s=await fetch(e,t);return s.ok&&(s.data=await s.json()),s};g.get=(n,e)=>g({url:n,...e});g.post=(n,e,t)=>g({method:"POST",url:n,body:JSON.stringify(e),...t});g.delete=(n,e)=>g({method:"DELETE",url:n,...e});g.put=(n,e,t)=>g({method:"PUT",url:n,body:JSON.stringify(e),...t});m(".first");m(".second");const N={shouldReject:!1,timeout:1e3,data:"아싸 성공!",errorMessage:"알 수 없는 오류가 발생했습니다."};function le(n){let e={...N};ce(n)&&(e.timeout=n),ae(n)&&(e={...N,...n});let{timeout:t,shouldReject:s,errorMessage:i,data:r}=e;return new Promise((o,a)=>{setTimeout(()=>{s?a({message:i}):o(r)},t)})}le(1e3);(function(){const n=m(".nav_scroll");window.addEventListener("scroll",()=>{window.scrollY>=190?(U(n,"display","flex"),ie(n,"sticky")):window.scrollY<190&&(U(n,"display","none"),ne(n,"sticky"))});const e=m(".top_banner"),t=m(".top_banner_close");function s(){e.style.transition="All 0.5s",e.style.transform="translateY(-50px)",setTimeout(()=>{e.remove()},250)}t.addEventListener("click",s)})();export{ue as a,U as c,m as g,pe as i,de as p,g as t};
