--- docs/assets/login-52VG-BlH.js
@@ -1,1 +1,0 @@
-import"./style-v06b81ne.js";import{a as s}from"./header-MICfn_ID.js";/* empty css               */import{g,s as f}from"./storage-jFd8YWAz.js";import{s as v}from"./setDocumentTitle-_OwlDZqT.js";import{p as h}from"./pocketbase-mqgmocw_.js";function m(t){return!!/^(?=.*[A-Za-z])[A-Za-z\d]{6,16}$/.test(t)}function p(t){return!!/^(?=.*[!@#$%^&*(),.?":{}|<>]).{6,16}$/.test(t)}const _=s("html"),o=s(".user_id_input"),l=s(".user_password_input"),d=s(".login_alert_container"),r=s(".error_message"),a=s("#user_password_error"),w=s(".login_btn");v("칼리 | 로그인");async function L(t,e){try{if(await h.collection("users").authWithPassword(t,e)){let{model:i,token:c}=await g("pocketbase_auth");f("auth",{isAuth:!!i,user:i,token:c}),window.location.href="/Karly/src/pages/main/"}}catch(n){u("visible","hidden"),console.log("로그인 실패!",n)}}function u(t,e){d.style.visibility=t,_.style.overflowY=e}d.addEventListener("click",function(t){t.target.tagName==="BUTTON"&&u("hidden","scroll")});w.addEventListener("click",function(t){t.preventDefault(),L(o.value,l.value)});o.addEventListener("input",t=>{let e=t.target.value;m(e)?r.classList.remove("is_invalid"):r.classList.add("is_invalid")});l.addEventListener("input",t=>{let e=t.target.value,n=p(e);e&&e.length<6?(a.classList.add("is_invalid"),a.textContent="최소 6자 이상 입력"):e.length>16?(a.classList.add("is_invalid"),a.textContent="최대 16자 까지 입력"):n?a.classList.remove("is_invalid"):a.textContent="특수문자 포함 최소 6자 - 최대 16자"});
