+++ docs/assets/main-tChaEllP.js
@@ -1,0 +1,1 @@
+import"./style-v06b81ne.js";import{g as be,c as R,p as he}from"./header-tMjJ4wC4.js";/* empty css               */import{r as Oe,a as ze,b as ke}from"./itemList-MN9rZj0T.js";import{c as De,d as Ge}from"./checkLogin-dd8Ug7jR.js";import{s as Be,g as Ve}from"./storage-PSdSBPgg.js";import{s as $e}from"./setDocumentTitle-_OwlDZqT.js";function ge(i){return i!==null&&typeof i=="object"&&"constructor"in i&&i.constructor===Object}function pe(i,e){i===void 0&&(i={}),e===void 0&&(e={}),Object.keys(e).forEach(t=>{typeof i[t]>"u"?i[t]=e[t]:ge(e[t])&&ge(i[t])&&Object.keys(e[t]).length>0&&pe(i[t],e[t])})}const xe={body:{},addEventListener(){},removeEventListener(){},activeElement:{blur(){},nodeName:""},querySelector(){return null},querySelectorAll(){return[]},getElementById(){return null},createEvent(){return{initEvent(){}}},createElement(){return{children:[],childNodes:[],style:{},setAttribute(){},getElementsByTagName(){return[]}}},createElementNS(){return{}},importNode(){return null},location:{hash:"",host:"",hostname:"",href:"",origin:"",pathname:"",protocol:"",search:""}};function N(){const i=typeof document<"u"?document:{};return pe(i,xe),i}const Fe={document:xe,navigator:{userAgent:""},location:{hash:"",host:"",hostname:"",href:"",origin:"",pathname:"",protocol:"",search:""},history:{replaceState(){},pushState(){},go(){},back(){}},CustomEvent:function(){return this},addEventListener(){},removeEventListener(){},getComputedStyle(){return{getPropertyValue(){return""}}},Image(){},Date(){},screen:{},setTimeout(){},clearTimeout(){},matchMedia(){return{}},requestAnimationFrame(i){return typeof setTimeout>"u"?(i(),null):setTimeout(i,0)},cancelAnimationFrame(i){typeof setTimeout>"u"||clearTimeout(i)}};function V(){const i=typeof window<"u"?window:{};return pe(i,Fe),i}function Ne(i){return i===void 0&&(i=""),i.trim().split(" ").filter(e=>!!e.trim())}function _e(i){const e=i;Object.keys(e).forEach(t=>{try{e[t]=null}catch{}try{delete e[t]}catch{}})}function ce(i,e){return e===void 0&&(e=0),setTimeout(i,e)}function J(){return Date.now()}function He(i){const e=V();let t;return e.getComputedStyle&&(t=e.getComputedStyle(i,null)),!t&&i.currentStyle&&(t=i.currentStyle),t||(t=i.style),t}function Re(i,e){e===void 0&&(e="x");const t=V();let s,n,r;const o=He(i);return t.WebKitCSSMatrix?(n=o.transform||o.webkitTransform,n.split(",").length>6&&(n=n.split(", ").map(l=>l.replace(",",".")).join(", ")),r=new t.WebKitCSSMatrix(n==="none"?"":n)):(r=o.MozTransform||o.OTransform||o.MsTransform||o.msTransform||o.transform||o.getPropertyValue("transform").replace("translate(","matrix(1, 0, 0, 1,"),s=r.toString().split(",")),e==="x"&&(t.WebKitCSSMatrix?n=r.m41:s.length===16?n=parseFloat(s[12]):n=parseFloat(s[4])),e==="y"&&(t.WebKitCSSMatrix?n=r.m42:s.length===16?n=parseFloat(s[13]):n=parseFloat(s[5])),n||0}function X(i){return typeof i=="object"&&i!==null&&i.constructor&&Object.prototype.toString.call(i).slice(8,-1)==="Object"}function We(i){return typeof window<"u"&&typeof window.HTMLElement<"u"?i instanceof HTMLElement:i&&(i.nodeType===1||i.nodeType===11)}function B(){const i=Object(arguments.length<=0?void 0:arguments[0]),e=["__proto__","constructor","prototype"];for(let t=1;t<arguments.length;t+=1){const s=t<0||arguments.length<=t?void 0:arguments[t];if(s!=null&&!We(s)){const n=Object.keys(Object(s)).filter(r=>e.indexOf(r)<0);for(let r=0,o=n.length;r<o;r+=1){const l=n[r],a=Object.getOwnPropertyDescriptor(s,l);a!==void 0&&a.enumerable&&(X(i[l])&&X(s[l])?s[l].__swiper__?i[l]=s[l]:B(i[l],s[l]):!X(i[l])&&X(s[l])?(i[l]={},s[l].__swiper__?i[l]=s[l]:B(i[l],s[l])):i[l]=s[l])}}}return i}function U(i,e,t){i.style.setProperty(e,t)}function Ee(i){let{swiper:e,targetPosition:t,side:s}=i;const n=V(),r=-e.translate;let o=null,l;const a=e.params.speed;e.wrapperEl.style.scrollSnapType="none",n.cancelAnimationFrame(e.cssModeFrameID);const f=t>r?"next":"prev",u=(h,m)=>f==="next"&&h>=m||f==="prev"&&h<=m,p=()=>{l=new Date().getTime(),o===null&&(o=l);const h=Math.max(Math.min((l-o)/a,1),0),m=.5-Math.cos(h*Math.PI)/2;let y=r+m*(t-r);if(u(y,t)&&(y=t),e.wrapperEl.scrollTo({[s]:y}),u(y,t)){e.wrapperEl.style.overflow="hidden",e.wrapperEl.style.scrollSnapType="",setTimeout(()=>{e.wrapperEl.style.overflow="",e.wrapperEl.scrollTo({[s]:y})}),n.cancelAnimationFrame(e.cssModeFrameID);return}e.cssModeFrameID=n.requestAnimationFrame(p)};p()}function F(i,e){return e===void 0&&(e=""),[...i.children].filter(t=>t.matches(e))}function Q(i){try{console.warn(i);return}catch{}}function Z(i,e){e===void 0&&(e=[]);const t=document.createElement(i);return t.classList.add(...Array.isArray(e)?e:Ne(e)),t}function je(i,e){const t=[];for(;i.previousElementSibling;){const s=i.previousElementSibling;e?s.matches(e)&&t.push(s):t.push(s),i=s}return t}function qe(i,e){const t=[];for(;i.nextElementSibling;){const s=i.nextElementSibling;e?s.matches(e)&&t.push(s):t.push(s),i=s}return t}function _(i,e){return V().getComputedStyle(i,null).getPropertyValue(e)}function ee(i){let e=i,t;if(e){for(t=0;(e=e.previousSibling)!==null;)e.nodeType===1&&(t+=1);return t}}function Ce(i,e){const t=[];let s=i.parentElement;for(;s;)e?s.matches(e)&&t.push(s):t.push(s),s=s.parentElement;return t}function ue(i,e,t){const s=V();return t?i[e==="width"?"offsetWidth":"offsetHeight"]+parseFloat(s.getComputedStyle(i,null).getPropertyValue(e==="width"?"margin-right":"margin-top"))+parseFloat(s.getComputedStyle(i,null).getPropertyValue(e==="width"?"margin-left":"margin-bottom")):i.offsetWidth}let re;function Ye(){const i=V(),e=N();return{smoothScroll:e.documentElement&&e.documentElement.style&&"scrollBehavior"in e.documentElement.style,touch:!!("ontouchstart"in i||i.DocumentTouch&&e instanceof i.DocumentTouch)}}function Pe(){return re||(re=Ye()),re}let ne;function Xe(i){let{userAgent:e}=i===void 0?{}:i;const t=Pe(),s=V(),n=s.navigator.platform,r=e||s.navigator.userAgent,o={ios:!1,android:!1},l=s.screen.width,a=s.screen.height,f=r.match(/(Android);?[\s\/]+([\d.]+)?/);let u=r.match(/(iPad).*OS\s([\d_]+)/);const p=r.match(/(iPod)(.*OS\s([\d_]+))?/),h=!u&&r.match(/(iPhone\sOS|iOS)\s([\d_]+)/),m=n==="Win32";let y=n==="MacIntel";const g=["1024x1366","1366x1024","834x1194","1194x834","834x1112","1112x834","768x1024","1024x768","820x1180","1180x820","810x1080","1080x810"];return!u&&y&&t.touch&&g.indexOf(`${l}x${a}`)>=0&&(u=r.match(/(Version)\/([\d.]+)/),u||(u=[0,1,"13_0_0"]),y=!1),f&&!m&&(o.os="android",o.android=!0),(u||h||p)&&(o.os="ios",o.ios=!0),o}function Ue(i){return i===void 0&&(i={}),ne||(ne=Xe(i)),ne}let ae;function Ke(){const i=V();let e=!1;function t(){const s=i.navigator.userAgent.toLowerCase();return s.indexOf("safari")>=0&&s.indexOf("chrome")<0&&s.indexOf("android")<0}if(t()){const s=String(i.navigator.userAgent);if(s.includes("Version/")){const[n,r]=s.split("Version/")[1].split(" ")[0].split(".").map(o=>Number(o));e=n<16||n===16&&r<2}}return{isSafari:e||t(),needPerspectiveFix:e,isWebView:/(iPhone|iPod|iPad).*AppleWebKit(?!.*Safari)/i.test(i.navigator.userAgent)}}function Je(){return ae||(ae=Ke()),ae}function Qe(i){let{swiper:e,on:t,emit:s}=i;const n=V();let r=null,o=null;const l=()=>{!e||e.destroyed||!e.initialized||(s("beforeResize"),s("resize"))},a=()=>{!e||e.destroyed||!e.initialized||(r=new ResizeObserver(p=>{o=n.requestAnimationFrame(()=>{const{width:h,height:m}=e;let y=h,g=m;p.forEach(S=>{let{contentBoxSize:w,contentRect:d,target:c}=S;c&&c!==e.el||(y=d?d.width:(w[0]||w).inlineSize,g=d?d.height:(w[0]||w).blockSize)}),(y!==h||g!==m)&&l()})}),r.observe(e.el))},f=()=>{o&&n.cancelAnimationFrame(o),r&&r.unobserve&&e.el&&(r.unobserve(e.el),r=null)},u=()=>{!e||e.destroyed||!e.initialized||s("orientationchange")};t("init",()=>{if(e.params.resizeObserver&&typeof n.ResizeObserver<"u"){a();return}n.addEventListener("resize",l),n.addEventListener("orientationchange",u)}),t("destroy",()=>{f(),n.removeEventListener("resize",l),n.removeEventListener("orientationchange",u)})}function Ze(i){let{swiper:e,extendParams:t,on:s,emit:n}=i;const r=[],o=V(),l=function(u,p){p===void 0&&(p={});const h=o.MutationObserver||o.WebkitMutationObserver,m=new h(y=>{if(e.__preventObserver__)return;if(y.length===1){n("observerUpdate",y[0]);return}const g=function(){n("observerUpdate",y[0])};o.requestAnimationFrame?o.requestAnimationFrame(g):o.setTimeout(g,0)});m.observe(u,{attributes:typeof p.attributes>"u"?!0:p.attributes,childList:typeof p.childList>"u"?!0:p.childList,characterData:typeof p.characterData>"u"?!0:p.characterData}),r.push(m)},a=()=>{if(e.params.observer){if(e.params.observeParents){const u=Ce(e.hostEl);for(let p=0;p<u.length;p+=1)l(u[p])}l(e.hostEl,{childList:e.params.observeSlideChildren}),l(e.wrapperEl,{attributes:!1})}},f=()=>{r.forEach(u=>{u.disconnect()}),r.splice(0,r.length)};t({observer:!1,observeParents:!1,observeSlideChildren:!1}),s("init",a),s("destroy",f)}var et={on(i,e,t){const s=this;if(!s.eventsListeners||s.destroyed||typeof e!="function")return s;const n=t?"unshift":"push";return i.split(" ").forEach(r=>{s.eventsListeners[r]||(s.eventsListeners[r]=[]),s.eventsListeners[r][n](e)}),s},once(i,e,t){const s=this;if(!s.eventsListeners||s.destroyed||typeof e!="function")return s;function n(){s.off(i,n),n.__emitterProxy&&delete n.__emitterProxy;for(var r=arguments.length,o=new Array(r),l=0;l<r;l++)o[l]=arguments[l];e.apply(s,o)}return n.__emitterProxy=e,s.on(i,n,t)},onAny(i,e){const t=this;if(!t.eventsListeners||t.destroyed||typeof i!="function")return t;const s=e?"unshift":"push";return t.eventsAnyListeners.indexOf(i)<0&&t.eventsAnyListeners[s](i),t},offAny(i){const e=this;if(!e.eventsListeners||e.destroyed||!e.eventsAnyListeners)return e;const t=e.eventsAnyListeners.indexOf(i);return t>=0&&e.eventsAnyListeners.splice(t,1),e},off(i,e){const t=this;return!t.eventsListeners||t.destroyed||!t.eventsListeners||i.split(" ").forEach(s=>{typeof e>"u"?t.eventsListeners[s]=[]:t.eventsListeners[s]&&t.eventsListeners[s].forEach((n,r)=>{(n===e||n.__emitterProxy&&n.__emitterProxy===e)&&t.eventsListeners[s].splice(r,1)})}),t},emit(){const i=this;if(!i.eventsListeners||i.destroyed||!i.eventsListeners)return i;let e,t,s;for(var n=arguments.length,r=new Array(n),o=0;o<n;o++)r[o]=arguments[o];return typeof r[0]=="string"||Array.isArray(r[0])?(e=r[0],t=r.slice(1,r.length),s=i):(e=r[0].events,t=r[0].data,s=r[0].context||i),t.unshift(s),(Array.isArray(e)?e:e.split(" ")).forEach(a=>{i.eventsAnyListeners&&i.eventsAnyListeners.length&&i.eventsAnyListeners.forEach(f=>{f.apply(s,[a,...t])}),i.eventsListeners&&i.eventsListeners[a]&&i.eventsListeners[a].forEach(f=>{f.apply(s,t)})}),i}};function tt(){const i=this;let e,t;const s=i.el;typeof i.params.width<"u"&&i.params.width!==null?e=i.params.width:e=s.clientWidth,typeof i.params.height<"u"&&i.params.height!==null?t=i.params.height:t=s.clientHeight,!(e===0&&i.isHorizontal()||t===0&&i.isVertical())&&(e=e-parseInt(_(s,"padding-left")||0,10)-parseInt(_(s,"padding-right")||0,10),t=t-parseInt(_(s,"padding-top")||0,10)-parseInt(_(s,"padding-bottom")||0,10),Number.isNaN(e)&&(e=0),Number.isNaN(t)&&(t=0),Object.assign(i,{width:e,height:t,size:i.isHorizontal()?e:t}))}function it(){const i=this;function e(b,E){return parseFloat(b.getPropertyValue(i.getDirectionLabel(E))||0)}const t=i.params,{wrapperEl:s,slidesEl:n,size:r,rtlTranslate:o,wrongRTL:l}=i,a=i.virtual&&t.virtual.enabled,f=a?i.virtual.slides.length:i.slides.length,u=F(n,`.${i.params.slideClass}, swiper-slide`),p=a?i.virtual.slides.length:u.length;let h=[];const m=[],y=[];let g=t.slidesOffsetBefore;typeof g=="function"&&(g=t.slidesOffsetBefore.call(i));let S=t.slidesOffsetAfter;typeof S=="function"&&(S=t.slidesOffsetAfter.call(i));const w=i.snapGrid.length,d=i.slidesGrid.length;let c=t.spaceBetween,v=-g,T=0,C=0;if(typeof r>"u")return;typeof c=="string"&&c.indexOf("%")>=0?c=parseFloat(c.replace("%",""))/100*r:typeof c=="string"&&(c=parseFloat(c)),i.virtualSize=-c,u.forEach(b=>{o?b.style.marginLeft="":b.style.marginRight="",b.style.marginBottom="",b.style.marginTop=""}),t.centeredSlides&&t.cssMode&&(U(s,"--swiper-centered-offset-before",""),U(s,"--swiper-centered-offset-after",""));const z=t.grid&&t.grid.rows>1&&i.grid;z?i.grid.initSlides(u):i.grid&&i.grid.unsetSlides();let A;const L=t.slidesPerView==="auto"&&t.breakpoints&&Object.keys(t.breakpoints).filter(b=>typeof t.breakpoints[b].slidesPerView<"u").length>0;for(let b=0;b<p;b+=1){A=0;let E;if(u[b]&&(E=u[b]),z&&i.grid.updateSlide(b,E,u),!(u[b]&&_(E,"display")==="none")){if(t.slidesPerView==="auto"){L&&(u[b].style[i.getDirectionLabel("width")]="");const P=getComputedStyle(E),x=E.style.transform,M=E.style.webkitTransform;if(x&&(E.style.transform="none"),M&&(E.style.webkitTransform="none"),t.roundLengths)A=i.isHorizontal()?ue(E,"width",!0):ue(E,"height",!0);else{const I=e(P,"width"),k=e(P,"padding-left"),se=e(P,"padding-right"),O=e(P,"margin-left"),$=e(P,"margin-right"),D=P.getPropertyValue("box-sizing");if(D&&D==="border-box")A=I+O+$;else{const{clientWidth:H,offsetWidth:Y}=E;A=I+k+se+O+$+(Y-H)}}x&&(E.style.transform=x),M&&(E.style.webkitTransform=M),t.roundLengths&&(A=Math.floor(A))}else A=(r-(t.slidesPerView-1)*c)/t.slidesPerView,t.roundLengths&&(A=Math.floor(A)),u[b]&&(u[b].style[i.getDirectionLabel("width")]=`${A}px`);u[b]&&(u[b].swiperSlideSize=A),y.push(A),t.centeredSlides?(v=v+A/2+T/2+c,T===0&&b!==0&&(v=v-r/2-c),b===0&&(v=v-r/2-c),Math.abs(v)<1/1e3&&(v=0),t.roundLengths&&(v=Math.floor(v)),C%t.slidesPerGroup===0&&h.push(v),m.push(v)):(t.roundLengths&&(v=Math.floor(v)),(C-Math.min(i.params.slidesPerGroupSkip,C))%i.params.slidesPerGroup===0&&h.push(v),m.push(v),v=v+A+c),i.virtualSize+=A+c,T=A,C+=1}}if(i.virtualSize=Math.max(i.virtualSize,r)+S,o&&l&&(t.effect==="slide"||t.effect==="coverflow")&&(s.style.width=`${i.virtualSize+c}px`),t.setWrapperSize&&(s.style[i.getDirectionLabel("width")]=`${i.virtualSize+c}px`),z&&i.grid.updateWrapperSize(A,h),!t.centeredSlides){const b=[];for(let E=0;E<h.length;E+=1){let P=h[E];t.roundLengths&&(P=Math.floor(P)),h[E]<=i.virtualSize-r&&b.push(P)}h=b,Math.floor(i.virtualSize-r)-Math.floor(h[h.length-1])>1&&h.push(i.virtualSize-r)}if(a&&t.loop){const b=y[0]+c;if(t.slidesPerGroup>1){const E=Math.ceil((i.virtual.slidesBefore+i.virtual.slidesAfter)/t.slidesPerGroup),P=b*t.slidesPerGroup;for(let x=0;x<E;x+=1)h.push(h[h.length-1]+P)}for(let E=0;E<i.virtual.slidesBefore+i.virtual.slidesAfter;E+=1)t.slidesPerGroup===1&&h.push(h[h.length-1]+b),m.push(m[m.length-1]+b),i.virtualSize+=b}if(h.length===0&&(h=[0]),c!==0){const b=i.isHorizontal()&&o?"marginLeft":i.getDirectionLabel("marginRight");u.filter((E,P)=>!t.cssMode||t.loop?!0:P!==u.length-1).forEach(E=>{E.style[b]=`${c}px`})}if(t.centeredSlides&&t.centeredSlidesBounds){let b=0;y.forEach(P=>{b+=P+(c||0)}),b-=c;const E=b-r;h=h.map(P=>P<=0?-g:P>E?E+S:P)}if(t.centerInsufficientSlides){let b=0;if(y.forEach(E=>{b+=E+(c||0)}),b-=c,b<r){const E=(r-b)/2;h.forEach((P,x)=>{h[x]=P-E}),m.forEach((P,x)=>{m[x]=P+E})}}if(Object.assign(i,{slides:u,snapGrid:h,slidesGrid:m,slidesSizesGrid:y}),t.centeredSlides&&t.cssMode&&!t.centeredSlidesBounds){U(s,"--swiper-centered-offset-before",`${-h[0]}px`),U(s,"--swiper-centered-offset-after",`${i.size/2-y[y.length-1]/2}px`);const b=-i.snapGrid[0],E=-i.slidesGrid[0];i.snapGrid=i.snapGrid.map(P=>P+b),i.slidesGrid=i.slidesGrid.map(P=>P+E)}if(p!==f&&i.emit("slidesLengthChange"),h.length!==w&&(i.params.watchOverflow&&i.checkOverflow(),i.emit("snapGridLengthChange")),m.length!==d&&i.emit("slidesGridLengthChange"),t.watchSlidesProgress&&i.updateSlidesOffset(),i.emit("slidesUpdated"),!a&&!t.cssMode&&(t.effect==="slide"||t.effect==="fade")){const b=`${t.containerModifierClass}backface-hidden`,E=i.el.classList.contains(b);p<=t.maxBackfaceHiddenSlides?E||i.el.classList.add(b):E&&i.el.classList.remove(b)}}function st(i){const e=this,t=[],s=e.virtual&&e.params.virtual.enabled;let n=0,r;typeof i=="number"?e.setTransition(i):i===!0&&e.setTransition(e.params.speed);const o=l=>s?e.slides[e.getSlideIndexByData(l)]:e.slides[l];if(e.params.slidesPerView!=="auto"&&e.params.slidesPerView>1)if(e.params.centeredSlides)(e.visibleSlides||[]).forEach(l=>{t.push(l)});else for(r=0;r<Math.ceil(e.params.slidesPerView);r+=1){const l=e.activeIndex+r;if(l>e.slides.length&&!s)break;t.push(o(l))}else t.push(o(e.activeIndex));for(r=0;r<t.length;r+=1)if(typeof t[r]<"u"){const l=t[r].offsetHeight;n=l>n?l:n}(n||n===0)&&(e.wrapperEl.style.height=`${n}px`)}function rt(){const i=this,e=i.slides,t=i.isElement?i.isHorizontal()?i.wrapperEl.offsetLeft:i.wrapperEl.offsetTop:0;for(let s=0;s<e.length;s+=1)e[s].swiperSlideOffset=(i.isHorizontal()?e[s].offsetLeft:e[s].offsetTop)-t-i.cssOverflowAdjustment()}function nt(i){i===void 0&&(i=this&&this.translate||0);const e=this,t=e.params,{slides:s,rtlTranslate:n,snapGrid:r}=e;if(s.length===0)return;typeof s[0].swiperSlideOffset>"u"&&e.updateSlidesOffset();let o=-i;n&&(o=i),s.forEach(a=>{a.classList.remove(t.slideVisibleClass,t.slideFullyVisibleClass)}),e.visibleSlidesIndexes=[],e.visibleSlides=[];let l=t.spaceBetween;typeof l=="string"&&l.indexOf("%")>=0?l=parseFloat(l.replace("%",""))/100*e.size:typeof l=="string"&&(l=parseFloat(l));for(let a=0;a<s.length;a+=1){const f=s[a];let u=f.swiperSlideOffset;t.cssMode&&t.centeredSlides&&(u-=s[0].swiperSlideOffset);const p=(o+(t.centeredSlides?e.minTranslate():0)-u)/(f.swiperSlideSize+l),h=(o-r[0]+(t.centeredSlides?e.minTranslate():0)-u)/(f.swiperSlideSize+l),m=-(o-u),y=m+e.slidesSizesGrid[a],g=m>=0&&m<=e.size-e.slidesSizesGrid[a];(m>=0&&m<e.size-1||y>1&&y<=e.size||m<=0&&y>=e.size)&&(e.visibleSlides.push(f),e.visibleSlidesIndexes.push(a),s[a].classList.add(t.slideVisibleClass)),g&&s[a].classList.add(t.slideFullyVisibleClass),f.progress=n?-p:p,f.originalProgress=n?-h:h}}function at(i){const e=this;if(typeof i>"u"){const u=e.rtlTranslate?-1:1;i=e&&e.translate&&e.translate*u||0}const t=e.params,s=e.maxTranslate()-e.minTranslate();let{progress:n,isBeginning:r,isEnd:o,progressLoop:l}=e;const a=r,f=o;if(s===0)n=0,r=!0,o=!0;else{n=(i-e.minTranslate())/s;const u=Math.abs(i-e.minTranslate())<1,p=Math.abs(i-e.maxTranslate())<1;r=u||n<=0,o=p||n>=1,u&&(n=0),p&&(n=1)}if(t.loop){const u=e.getSlideIndexByData(0),p=e.getSlideIndexByData(e.slides.length-1),h=e.slidesGrid[u],m=e.slidesGrid[p],y=e.slidesGrid[e.slidesGrid.length-1],g=Math.abs(i);g>=h?l=(g-h)/y:l=(g+y-m)/y,l>1&&(l-=1)}Object.assign(e,{progress:n,progressLoop:l,isBeginning:r,isEnd:o}),(t.watchSlidesProgress||t.centeredSlides&&t.autoHeight)&&e.updateSlidesProgress(i),r&&!a&&e.emit("reachBeginning toEdge"),o&&!f&&e.emit("reachEnd toEdge"),(a&&!r||f&&!o)&&e.emit("fromEdge"),e.emit("progress",n)}function lt(){const i=this,{slides:e,params:t,slidesEl:s,activeIndex:n}=i,r=i.virtual&&t.virtual.enabled,o=i.grid&&t.grid&&t.grid.rows>1,l=p=>F(s,`.${t.slideClass}${p}, swiper-slide${p}`)[0];e.forEach(p=>{p.classList.remove(t.slideActiveClass,t.slideNextClass,t.slidePrevClass)});let a,f,u;if(r)if(t.loop){let p=n-i.virtual.slidesBefore;p<0&&(p=i.virtual.slides.length+p),p>=i.virtual.slides.length&&(p-=i.virtual.slides.length),a=l(`[data-swiper-slide-index="${p}"]`)}else a=l(`[data-swiper-slide-index="${n}"]`);else o?(a=e.filter(p=>p.column===n)[0],u=e.filter(p=>p.column===n+1)[0],f=e.filter(p=>p.column===n-1)[0]):a=e[n];a&&(a.classList.add(t.slideActiveClass),o?(u&&u.classList.add(t.slideNextClass),f&&f.classList.add(t.slidePrevClass)):(u=qe(a,`.${t.slideClass}, swiper-slide`)[0],t.loop&&!u&&(u=e[0]),u&&u.classList.add(t.slideNextClass),f=je(a,`.${t.slideClass}, swiper-slide`)[0],t.loop&&!f===0&&(f=e[e.length-1]),f&&f.classList.add(t.slidePrevClass))),i.emitSlidesClasses()}const K=(i,e)=>{if(!i||i.destroyed||!i.params)return;const t=()=>i.isElement?"swiper-slide":`.${i.params.slideClass}`,s=e.closest(t());if(s){let n=s.querySelector(`.${i.params.lazyPreloaderClass}`);!n&&i.isElement&&(s.shadowRoot?n=s.shadowRoot.querySelector(`.${i.params.lazyPreloaderClass}`):requestAnimationFrame(()=>{s.shadowRoot&&(n=s.shadowRoot.querySelector(`.${i.params.lazyPreloaderClass}`),n&&n.remove())})),n&&n.remove()}},le=(i,e)=>{if(!i.slides[e])return;const t=i.slides[e].querySelector('[loading="lazy"]');t&&t.removeAttribute("loading")},fe=i=>{if(!i||i.destroyed||!i.params)return;let e=i.params.lazyPreloadPrevNext;const t=i.slides.length;if(!t||!e||e<0)return;e=Math.min(e,t);const s=i.params.slidesPerView==="auto"?i.slidesPerViewDynamic():Math.ceil(i.params.slidesPerView),n=i.activeIndex;if(i.params.grid&&i.params.grid.rows>1){const o=n,l=[o-e];l.push(...Array.from({length:e}).map((a,f)=>o+s+f)),i.slides.forEach((a,f)=>{l.includes(a.column)&&le(i,f)});return}const r=n+s-1;if(i.params.rewind||i.params.loop)for(let o=n-e;o<=r+e;o+=1){const l=(o%t+t)%t;(l<n||l>r)&&le(i,l)}else for(let o=Math.max(n-e,0);o<=Math.min(r+e,t-1);o+=1)o!==n&&(o>r||o<n)&&le(i,o)};function ot(i){const{slidesGrid:e,params:t}=i,s=i.rtlTranslate?i.translate:-i.translate;let n;for(let r=0;r<e.length;r+=1)typeof e[r+1]<"u"?s>=e[r]&&s<e[r+1]-(e[r+1]-e[r])/2?n=r:s>=e[r]&&s<e[r+1]&&(n=r+1):s>=e[r]&&(n=r);return t.normalizeSlideIndex&&(n<0||typeof n>"u")&&(n=0),n}function dt(i){const e=this,t=e.rtlTranslate?e.translate:-e.translate,{snapGrid:s,params:n,activeIndex:r,realIndex:o,snapIndex:l}=e;let a=i,f;const u=m=>{let y=m-e.virtual.slidesBefore;return y<0&&(y=e.virtual.slides.length+y),y>=e.virtual.slides.length&&(y-=e.virtual.slides.length),y};if(typeof a>"u"&&(a=ot(e)),s.indexOf(t)>=0)f=s.indexOf(t);else{const m=Math.min(n.slidesPerGroupSkip,a);f=m+Math.floor((a-m)/n.slidesPerGroup)}if(f>=s.length&&(f=s.length-1),a===r&&!e.params.loop){f!==l&&(e.snapIndex=f,e.emit("snapIndexChange"));return}if(a===r&&e.params.loop&&e.virtual&&e.params.virtual.enabled){e.realIndex=u(a);return}const p=e.grid&&n.grid&&n.grid.rows>1;let h;if(e.virtual&&n.virtual.enabled&&n.loop)h=u(a);else if(p){const m=e.slides.filter(g=>g.column===a)[0];let y=parseInt(m.getAttribute("data-swiper-slide-index"),10);Number.isNaN(y)&&(y=Math.max(e.slides.indexOf(m),0)),h=Math.floor(y/n.grid.rows)}else if(e.slides[a]){const m=e.slides[a].getAttribute("data-swiper-slide-index");m?h=parseInt(m,10):h=a}else h=a;Object.assign(e,{previousSnapIndex:l,snapIndex:f,previousRealIndex:o,realIndex:h,previousIndex:r,activeIndex:a}),e.initialized&&fe(e),e.emit("activeIndexChange"),e.emit("snapIndexChange"),(e.initialized||e.params.runCallbacksOnInit)&&(o!==h&&e.emit("realIndexChange"),e.emit("slideChange"))}function ct(i,e){const t=this,s=t.params;let n=i.closest(`.${s.slideClass}, swiper-slide`);!n&&t.isElement&&e&&e.length>1&&e.includes(i)&&[...e.slice(e.indexOf(i)+1,e.length)].forEach(l=>{!n&&l.matches&&l.matches(`.${s.slideClass}, swiper-slide`)&&(n=l)});let r=!1,o;if(n){for(let l=0;l<t.slides.length;l+=1)if(t.slides[l]===n){r=!0,o=l;break}}if(n&&r)t.clickedSlide=n,t.virtual&&t.params.virtual.enabled?t.clickedIndex=parseInt(n.getAttribute("data-swiper-slide-index"),10):t.clickedIndex=o;else{t.clickedSlide=void 0,t.clickedIndex=void 0;return}s.slideToClickedSlide&&t.clickedIndex!==void 0&&t.clickedIndex!==t.activeIndex&&t.slideToClickedSlide()}var ut={updateSize:tt,updateSlides:it,updateAutoHeight:st,updateSlidesOffset:rt,updateSlidesProgress:nt,updateProgress:at,updateSlidesClasses:lt,updateActiveIndex:dt,updateClickedSlide:ct};function ft(i){i===void 0&&(i=this.isHorizontal()?"x":"y");const e=this,{params:t,rtlTranslate:s,translate:n,wrapperEl:r}=e;if(t.virtualTranslate)return s?-n:n;if(t.cssMode)return n;let o=Re(r,i);return o+=e.cssOverflowAdjustment(),s&&(o=-o),o||0}function pt(i,e){const t=this,{rtlTranslate:s,params:n,wrapperEl:r,progress:o}=t;let l=0,a=0;const f=0;t.isHorizontal()?l=s?-i:i:a=i,n.roundLengths&&(l=Math.floor(l),a=Math.floor(a)),t.previousTranslate=t.translate,t.translate=t.isHorizontal()?l:a,n.cssMode?r[t.isHorizontal()?"scrollLeft":"scrollTop"]=t.isHorizontal()?-l:-a:n.virtualTranslate||(t.isHorizontal()?l-=t.cssOverflowAdjustment():a-=t.cssOverflowAdjustment(),r.style.transform=`translate3d(${l}px, ${a}px, ${f}px)`);let u;const p=t.maxTranslate()-t.minTranslate();p===0?u=0:u=(i-t.minTranslate())/p,u!==o&&t.updateProgress(i),t.emit("setTranslate",t.translate,e)}function mt(){return-this.snapGrid[0]}function ht(){return-this.snapGrid[this.snapGrid.length-1]}function gt(i,e,t,s,n){i===void 0&&(i=0),e===void 0&&(e=this.params.speed),t===void 0&&(t=!0),s===void 0&&(s=!0);const r=this,{params:o,wrapperEl:l}=r;if(r.animating&&o.preventInteractionOnTransition)return!1;const a=r.minTranslate(),f=r.maxTranslate();let u;if(s&&i>a?u=a:s&&i<f?u=f:u=i,r.updateProgress(u),o.cssMode){const p=r.isHorizontal();if(e===0)l[p?"scrollLeft":"scrollTop"]=-u;else{if(!r.support.smoothScroll)return Ee({swiper:r,targetPosition:-u,side:p?"left":"top"}),!0;l.scrollTo({[p?"left":"top"]:-u,behavior:"smooth"})}return!0}return e===0?(r.setTransition(0),r.setTranslate(u),t&&(r.emit("beforeTransitionStart",e,n),r.emit("transitionEnd"))):(r.setTransition(e),r.setTranslate(u),t&&(r.emit("beforeTransitionStart",e,n),r.emit("transitionStart")),r.animating||(r.animating=!0,r.onTranslateToWrapperTransitionEnd||(r.onTranslateToWrapperTransitionEnd=function(h){!r||r.destroyed||h.target===this&&(r.wrapperEl.removeEventListener("transitionend",r.onTranslateToWrapperTransitionEnd),r.onTranslateToWrapperTransitionEnd=null,delete r.onTranslateToWrapperTransitionEnd,t&&r.emit("transitionEnd"))}),r.wrapperEl.addEventListener("transitionend",r.onTranslateToWrapperTransitionEnd))),!0}var vt={getTranslate:ft,setTranslate:pt,minTranslate:mt,maxTranslate:ht,translateTo:gt};function wt(i,e){const t=this;t.params.cssMode||(t.wrapperEl.style.transitionDuration=`${i}ms`,t.wrapperEl.style.transitionDelay=i===0?"0ms":""),t.emit("setTransition",i,e)}function Me(i){let{swiper:e,runCallbacks:t,direction:s,step:n}=i;const{activeIndex:r,previousIndex:o}=e;let l=s;if(l||(r>o?l="next":r<o?l="prev":l="reset"),e.emit(`transition${n}`),t&&r!==o){if(l==="reset"){e.emit(`slideResetTransition${n}`);return}e.emit(`slideChangeTransition${n}`),l==="next"?e.emit(`slideNextTransition${n}`):e.emit(`slidePrevTransition${n}`)}}function yt(i,e){i===void 0&&(i=!0);const t=this,{params:s}=t;s.cssMode||(s.autoHeight&&t.updateAutoHeight(),Me({swiper:t,runCallbacks:i,direction:e,step:"Start"}))}function St(i,e){i===void 0&&(i=!0);const t=this,{params:s}=t;t.animating=!1,!s.cssMode&&(t.setTransition(0),Me({swiper:t,runCallbacks:i,direction:e,step:"End"}))}var Tt={setTransition:wt,transitionStart:yt,transitionEnd:St};function bt(i,e,t,s,n){i===void 0&&(i=0),e===void 0&&(e=this.params.speed),t===void 0&&(t=!0),typeof i=="string"&&(i=parseInt(i,10));const r=this;let o=i;o<0&&(o=0);const{params:l,snapGrid:a,slidesGrid:f,previousIndex:u,activeIndex:p,rtlTranslate:h,wrapperEl:m,enabled:y}=r;if(r.animating&&l.preventInteractionOnTransition||!y&&!s&&!n)return!1;const g=Math.min(r.params.slidesPerGroupSkip,o);let S=g+Math.floor((o-g)/r.params.slidesPerGroup);S>=a.length&&(S=a.length-1);const w=-a[S];if(l.normalizeSlideIndex)for(let c=0;c<f.length;c+=1){const v=-Math.floor(w*100),T=Math.floor(f[c]*100),C=Math.floor(f[c+1]*100);typeof f[c+1]<"u"?v>=T&&v<C-(C-T)/2?o=c:v>=T&&v<C&&(o=c+1):v>=T&&(o=c)}if(r.initialized&&o!==p&&(!r.allowSlideNext&&(h?w>r.translate&&w>r.minTranslate():w<r.translate&&w<r.minTranslate())||!r.allowSlidePrev&&w>r.translate&&w>r.maxTranslate()&&(p||0)!==o))return!1;o!==(u||0)&&t&&r.emit("beforeSlideChangeStart"),r.updateProgress(w);let d;if(o>p?d="next":o<p?d="prev":d="reset",h&&-w===r.translate||!h&&w===r.translate)return r.updateActiveIndex(o),l.autoHeight&&r.updateAutoHeight(),r.updateSlidesClasses(),l.effect!=="slide"&&r.setTranslate(w),d!=="reset"&&(r.transitionStart(t,d),r.transitionEnd(t,d)),!1;if(l.cssMode){const c=r.isHorizontal(),v=h?w:-w;if(e===0){const T=r.virtual&&r.params.virtual.enabled;T&&(r.wrapperEl.style.scrollSnapType="none",r._immediateVirtual=!0),T&&!r._cssModeVirtualInitialSet&&r.params.initialSlide>0?(r._cssModeVirtualInitialSet=!0,requestAnimationFrame(()=>{m[c?"scrollLeft":"scrollTop"]=v})):m[c?"scrollLeft":"scrollTop"]=v,T&&requestAnimationFrame(()=>{r.wrapperEl.style.scrollSnapType="",r._immediateVirtual=!1})}else{if(!r.support.smoothScroll)return Ee({swiper:r,targetPosition:v,side:c?"left":"top"}),!0;m.scrollTo({[c?"left":"top"]:v,behavior:"smooth"})}return!0}return r.setTransition(e),r.setTranslate(w),r.updateActiveIndex(o),r.updateSlidesClasses(),r.emit("beforeTransitionStart",e,s),r.transitionStart(t,d),e===0?r.transitionEnd(t,d):r.animating||(r.animating=!0,r.onSlideToWrapperTransitionEnd||(r.onSlideToWrapperTransitionEnd=function(v){!r||r.destroyed||v.target===this&&(r.wrapperEl.removeEventListener("transitionend",r.onSlideToWrapperTransitionEnd),r.onSlideToWrapperTransitionEnd=null,delete r.onSlideToWrapperTransitionEnd,r.transitionEnd(t,d))}),r.wrapperEl.addEventListener("transitionend",r.onSlideToWrapperTransitionEnd)),!0}function xt(i,e,t,s){i===void 0&&(i=0),e===void 0&&(e=this.params.speed),t===void 0&&(t=!0),typeof i=="string"&&(i=parseInt(i,10));const n=this,r=n.grid&&n.params.grid&&n.params.grid.rows>1;let o=i;if(n.params.loop)if(n.virtual&&n.params.virtual.enabled)o=o+n.virtual.slidesBefore;else{let l;if(r){const h=o*n.params.grid.rows;l=n.slides.filter(m=>m.getAttribute("data-swiper-slide-index")*1===h)[0].column}else l=n.getSlideIndexByData(o);const a=r?Math.ceil(n.slides.length/n.params.grid.rows):n.slides.length,{centeredSlides:f}=n.params;let u=n.params.slidesPerView;u==="auto"?u=n.slidesPerViewDynamic():(u=Math.ceil(parseFloat(n.params.slidesPerView,10)),f&&u%2===0&&(u=u+1));let p=a-l<u;if(f&&(p=p||l<Math.ceil(u/2)),p){const h=f?l<n.activeIndex?"prev":"next":l-n.activeIndex-1<n.params.slidesPerView?"next":"prev";n.loopFix({direction:h,slideTo:!0,activeSlideIndex:h==="next"?l+1:l-a+1,slideRealIndex:h==="next"?n.realIndex:void 0})}if(r){const h=o*n.params.grid.rows;o=n.slides.filter(m=>m.getAttribute("data-swiper-slide-index")*1===h)[0].column}else o=n.getSlideIndexByData(o)}return requestAnimationFrame(()=>{n.slideTo(o,e,t,s)}),n}function Et(i,e,t){i===void 0&&(i=this.params.speed),e===void 0&&(e=!0);const s=this,{enabled:n,params:r,animating:o}=s;if(!n)return s;let l=r.slidesPerGroup;r.slidesPerView==="auto"&&r.slidesPerGroup===1&&r.slidesPerGroupAuto&&(l=Math.max(s.slidesPerViewDynamic("current",!0),1));const a=s.activeIndex<r.slidesPerGroupSkip?1:l,f=s.virtual&&r.virtual.enabled;if(r.loop){if(o&&!f&&r.loopPreventsSliding)return!1;if(s.loopFix({direction:"next"}),s._clientLeft=s.wrapperEl.clientLeft,s.activeIndex===s.slides.length-1&&r.cssMode)return requestAnimationFrame(()=>{s.slideTo(s.activeIndex+a,i,e,t)}),!0}return r.rewind&&s.isEnd?s.slideTo(0,i,e,t):s.slideTo(s.activeIndex+a,i,e,t)}function Ct(i,e,t){i===void 0&&(i=this.params.speed),e===void 0&&(e=!0);const s=this,{params:n,snapGrid:r,slidesGrid:o,rtlTranslate:l,enabled:a,animating:f}=s;if(!a)return s;const u=s.virtual&&n.virtual.enabled;if(n.loop){if(f&&!u&&n.loopPreventsSliding)return!1;s.loopFix({direction:"prev"}),s._clientLeft=s.wrapperEl.clientLeft}const p=l?s.translate:-s.translate;function h(w){return w<0?-Math.floor(Math.abs(w)):Math.floor(w)}const m=h(p),y=r.map(w=>h(w));let g=r[y.indexOf(m)-1];if(typeof g>"u"&&n.cssMode){let w;r.forEach((d,c)=>{m>=d&&(w=c)}),typeof w<"u"&&(g=r[w>0?w-1:w])}let S=0;if(typeof g<"u"&&(S=o.indexOf(g),S<0&&(S=s.activeIndex-1),n.slidesPerView==="auto"&&n.slidesPerGroup===1&&n.slidesPerGroupAuto&&(S=S-s.slidesPerViewDynamic("previous",!0)+1,S=Math.max(S,0))),n.rewind&&s.isBeginning){const w=s.params.virtual&&s.params.virtual.enabled&&s.virtual?s.virtual.slides.length-1:s.slides.length-1;return s.slideTo(w,i,e,t)}else if(n.loop&&s.activeIndex===0&&n.cssMode)return requestAnimationFrame(()=>{s.slideTo(S,i,e,t)}),!0;return s.slideTo(S,i,e,t)}function Pt(i,e,t){i===void 0&&(i=this.params.speed),e===void 0&&(e=!0);const s=this;return s.slideTo(s.activeIndex,i,e,t)}function Mt(i,e,t,s){i===void 0&&(i=this.params.speed),e===void 0&&(e=!0),s===void 0&&(s=.5);const n=this;let r=n.activeIndex;const o=Math.min(n.params.slidesPerGroupSkip,r),l=o+Math.floor((r-o)/n.params.slidesPerGroup),a=n.rtlTranslate?n.translate:-n.translate;if(a>=n.snapGrid[l]){const f=n.snapGrid[l],u=n.snapGrid[l+1];a-f>(u-f)*s&&(r+=n.params.slidesPerGroup)}else{const f=n.snapGrid[l-1],u=n.snapGrid[l];a-f<=(u-f)*s&&(r-=n.params.slidesPerGroup)}return r=Math.max(r,0),r=Math.min(r,n.slidesGrid.length-1),n.slideTo(r,i,e,t)}function Lt(){const i=this,{params:e,slidesEl:t}=i,s=e.slidesPerView==="auto"?i.slidesPerViewDynamic():e.slidesPerView;let n=i.clickedIndex,r;const o=i.isElement?"swiper-slide":`.${e.slideClass}`;if(e.loop){if(i.animating)return;r=parseInt(i.clickedSlide.getAttribute("data-swiper-slide-index"),10),e.centeredSlides?n<i.loopedSlides-s/2||n>i.slides.length-i.loopedSlides+s/2?(i.loopFix(),n=i.getSlideIndex(F(t,`${o}[data-swiper-slide-index="${r}"]`)[0]),ce(()=>{i.slideTo(n)})):i.slideTo(n):n>i.slides.length-s?(i.loopFix(),n=i.getSlideIndex(F(t,`${o}[data-swiper-slide-index="${r}"]`)[0]),ce(()=>{i.slideTo(n)})):i.slideTo(n)}else i.slideTo(n)}var It={slideTo:bt,slideToLoop:xt,slideNext:Et,slidePrev:Ct,slideReset:Pt,slideToClosest:Mt,slideToClickedSlide:Lt};function At(i){const e=this,{params:t,slidesEl:s}=e;if(!t.loop||e.virtual&&e.params.virtual.enabled)return;const n=()=>{F(s,`.${t.slideClass}, swiper-slide`).forEach((p,h)=>{p.setAttribute("data-swiper-slide-index",h)})},r=e.grid&&t.grid&&t.grid.rows>1,o=t.slidesPerGroup*(r?t.grid.rows:1),l=e.slides.length%o!==0,a=r&&e.slides.length%t.grid.rows!==0,f=u=>{for(let p=0;p<u;p+=1){const h=e.isElement?Z("swiper-slide",[t.slideBlankClass]):Z("div",[t.slideClass,t.slideBlankClass]);e.slidesEl.append(h)}};if(l){if(t.loopAddBlankSlides){const u=o-e.slides.length%o;f(u),e.recalcSlides(),e.updateSlides()}else Q("Swiper Loop Warning: The number of slides is not even to slidesPerGroup, loop mode may not function properly. You need to add more slides (or make duplicates, or empty slides)");n()}else if(a){if(t.loopAddBlankSlides){const u=t.grid.rows-e.slides.length%t.grid.rows;f(u),e.recalcSlides(),e.updateSlides()}else Q("Swiper Loop Warning: The number of slides is not even to grid.rows, loop mode may not function properly. You need to add more slides (or make duplicates, or empty slides)");n()}else n();e.loopFix({slideRealIndex:i,direction:t.centeredSlides?void 0:"next"})}function Ot(i){let{slideRealIndex:e,slideTo:t=!0,direction:s,setTranslate:n,activeSlideIndex:r,byController:o,byMousewheel:l}=i===void 0?{}:i;const a=this;if(!a.params.loop)return;a.emit("beforeLoopFix");const{slides:f,allowSlidePrev:u,allowSlideNext:p,slidesEl:h,params:m}=a,{centeredSlides:y}=m;if(a.allowSlidePrev=!0,a.allowSlideNext=!0,a.virtual&&m.virtual.enabled){t&&(!m.centeredSlides&&a.snapIndex===0?a.slideTo(a.virtual.slides.length,0,!1,!0):m.centeredSlides&&a.snapIndex<m.slidesPerView?a.slideTo(a.virtual.slides.length+a.snapIndex,0,!1,!0):a.snapIndex===a.snapGrid.length-1&&a.slideTo(a.virtual.slidesBefore,0,!1,!0)),a.allowSlidePrev=u,a.allowSlideNext=p,a.emit("loopFix");return}let g=m.slidesPerView;g==="auto"?g=a.slidesPerViewDynamic():(g=Math.ceil(parseFloat(m.slidesPerView,10)),y&&g%2===0&&(g=g+1));const S=m.slidesPerGroupAuto?g:m.slidesPerGroup;let w=S;w%S!==0&&(w+=S-w%S),w+=m.loopAdditionalSlides,a.loopedSlides=w;const d=a.grid&&m.grid&&m.grid.rows>1;f.length<g+w?Q("Swiper Loop Warning: The number of slides is not enough for loop mode, it will be disabled and not function properly. You need to add more slides (or make duplicates) or lower the values of slidesPerView and slidesPerGroup parameters"):d&&m.grid.fill==="row"&&Q("Swiper Loop Warning: Loop mode is not compatible with grid.fill = `row`");const c=[],v=[];let T=a.activeIndex;typeof r>"u"?r=a.getSlideIndex(f.filter(x=>x.classList.contains(m.slideActiveClass))[0]):T=r;const C=s==="next"||!s,z=s==="prev"||!s;let A=0,L=0;const b=d?Math.ceil(f.length/m.grid.rows):f.length,P=(d?f[r].column:r)+(y&&typeof n>"u"?-g/2+.5:0);if(P<w){A=Math.max(w-P,S);for(let x=0;x<w-P;x+=1){const M=x-Math.floor(x/b)*b;if(d){const I=b-M-1;for(let k=f.length-1;k>=0;k-=1)f[k].column===I&&c.push(k)}else c.push(b-M-1)}}else if(P+g>b-w){L=Math.max(P-(b-w*2),S);for(let x=0;x<L;x+=1){const M=x-Math.floor(x/b)*b;d?f.forEach((I,k)=>{I.column===M&&v.push(k)}):v.push(M)}}if(a.__preventObserver__=!0,requestAnimationFrame(()=>{a.__preventObserver__=!1}),z&&c.forEach(x=>{f[x].swiperLoopMoveDOM=!0,h.prepend(f[x]),f[x].swiperLoopMoveDOM=!1}),C&&v.forEach(x=>{f[x].swiperLoopMoveDOM=!0,h.append(f[x]),f[x].swiperLoopMoveDOM=!1}),a.recalcSlides(),m.slidesPerView==="auto"?a.updateSlides():d&&(c.length>0&&z||v.length>0&&C)&&a.slides.forEach((x,M)=>{a.grid.updateSlide(M,x,a.slides)}),m.watchSlidesProgress&&a.updateSlidesOffset(),t){if(c.length>0&&z){if(typeof e>"u"){const x=a.slidesGrid[T],I=a.slidesGrid[T+A]-x;l?a.setTranslate(a.translate-I):(a.slideTo(T+A,0,!1,!0),n&&(a.touchEventsData.startTranslate=a.touchEventsData.startTranslate-I,a.touchEventsData.currentTranslate=a.touchEventsData.currentTranslate-I))}else if(n){const x=d?c.length/m.grid.rows:c.length;a.slideTo(a.activeIndex+x,0,!1,!0),a.touchEventsData.currentTranslate=a.translate}}else if(v.length>0&&C)if(typeof e>"u"){const x=a.slidesGrid[T],I=a.slidesGrid[T-L]-x;l?a.setTranslate(a.translate-I):(a.slideTo(T-L,0,!1,!0),n&&(a.touchEventsData.startTranslate=a.touchEventsData.startTranslate-I,a.touchEventsData.currentTranslate=a.touchEventsData.currentTranslate-I))}else{const x=d?v.length/m.grid.rows:v.length;a.slideTo(a.activeIndex-x,0,!1,!0)}}if(a.allowSlidePrev=u,a.allowSlideNext=p,a.controller&&a.controller.control&&!o){const x={slideRealIndex:e,direction:s,setTranslate:n,activeSlideIndex:r,byController:!0};Array.isArray(a.controller.control)?a.controller.control.forEach(M=>{!M.destroyed&&M.params.loop&&M.loopFix({...x,slideTo:M.params.slidesPerView===m.slidesPerView?t:!1})}):a.controller.control instanceof a.constructor&&a.controller.control.params.loop&&a.controller.control.loopFix({...x,slideTo:a.controller.control.params.slidesPerView===m.slidesPerView?t:!1})}a.emit("loopFix")}function zt(){const i=this,{params:e,slidesEl:t}=i;if(!e.loop||i.virtual&&i.params.virtual.enabled)return;i.recalcSlides();const s=[];i.slides.forEach(n=>{const r=typeof n.swiperSlideIndex>"u"?n.getAttribute("data-swiper-slide-index")*1:n.swiperSlideIndex;s[r]=n}),i.slides.forEach(n=>{n.removeAttribute("data-swiper-slide-index")}),s.forEach(n=>{t.append(n)}),i.recalcSlides(),i.slideTo(i.realIndex,0)}var kt={loopCreate:At,loopFix:Ot,loopDestroy:zt};function Dt(i){const e=this;if(!e.params.simulateTouch||e.params.watchOverflow&&e.isLocked||e.params.cssMode)return;const t=e.params.touchEventsTarget==="container"?e.el:e.wrapperEl;e.isElement&&(e.__preventObserver__=!0),t.style.cursor="move",t.style.cursor=i?"grabbing":"grab",e.isElement&&requestAnimationFrame(()=>{e.__preventObserver__=!1})}function Gt(){const i=this;i.params.watchOverflow&&i.isLocked||i.params.cssMode||(i.isElement&&(i.__preventObserver__=!0),i[i.params.touchEventsTarget==="container"?"el":"wrapperEl"].style.cursor="",i.isElement&&requestAnimationFrame(()=>{i.__preventObserver__=!1}))}var Bt={setGrabCursor:Dt,unsetGrabCursor:Gt};function Vt(i,e){e===void 0&&(e=this);function t(s){if(!s||s===N()||s===V())return null;s.assignedSlot&&(s=s.assignedSlot);const n=s.closest(i);return!n&&!s.getRootNode?null:n||t(s.getRootNode().host)}return t(e)}function ve(i,e,t){const s=V(),{params:n}=i,r=n.edgeSwipeDetection,o=n.edgeSwipeThreshold;return r&&(t<=o||t>=s.innerWidth-o)?r==="prevent"?(e.preventDefault(),!0):!1:!0}function $t(i){const e=this,t=N();let s=i;s.originalEvent&&(s=s.originalEvent);const n=e.touchEventsData;if(s.type==="pointerdown"){if(n.pointerId!==null&&n.pointerId!==s.pointerId)return;n.pointerId=s.pointerId}else s.type==="touchstart"&&s.targetTouches.length===1&&(n.touchId=s.targetTouches[0].identifier);if(s.type==="touchstart"){ve(e,s,s.targetTouches[0].pageX);return}const{params:r,touches:o,enabled:l}=e;if(!l||!r.simulateTouch&&s.pointerType==="mouse"||e.animating&&r.preventInteractionOnTransition)return;!e.animating&&r.cssMode&&r.loop&&e.loopFix();let a=s.target;if(r.touchEventsTarget==="wrapper"&&!e.wrapperEl.contains(a)||"which"in s&&s.which===3||"button"in s&&s.button>0||n.isTouched&&n.isMoved)return;const f=!!r.noSwipingClass&&r.noSwipingClass!=="",u=s.composedPath?s.composedPath():s.path;f&&s.target&&s.target.shadowRoot&&u&&(a=u[0]);const p=r.noSwipingSelector?r.noSwipingSelector:`.${r.noSwipingClass}`,h=!!(s.target&&s.target.shadowRoot);if(r.noSwiping&&(h?Vt(p,a):a.closest(p))){e.allowClick=!0;return}if(r.swipeHandler&&!a.closest(r.swipeHandler))return;o.currentX=s.pageX,o.currentY=s.pageY;const m=o.currentX,y=o.currentY;if(!ve(e,s,m))return;Object.assign(n,{isTouched:!0,isMoved:!1,allowTouchCallbacks:!0,isScrolling:void 0,startMoving:void 0}),o.startX=m,o.startY=y,n.touchStartTime=J(),e.allowClick=!0,e.updateSize(),e.swipeDirection=void 0,r.threshold>0&&(n.allowThresholdMove=!1);let g=!0;a.matches(n.focusableElements)&&(g=!1,a.nodeName==="SELECT"&&(n.isTouched=!1)),t.activeElement&&t.activeElement.matches(n.focusableElements)&&t.activeElement!==a&&t.activeElement.blur();const S=g&&e.allowTouchMove&&r.touchStartPreventDefault;(r.touchStartForcePreventDefault||S)&&!a.isContentEditable&&s.preventDefault(),r.freeMode&&r.freeMode.enabled&&e.freeMode&&e.animating&&!r.cssMode&&e.freeMode.onTouchStart(),e.emit("touchStart",s)}function Ft(i){const e=N(),t=this,s=t.touchEventsData,{params:n,touches:r,rtlTranslate:o,enabled:l}=t;if(!l||!n.simulateTouch&&i.pointerType==="mouse")return;let a=i;if(a.originalEvent&&(a=a.originalEvent),a.type==="pointermove"&&(s.touchId!==null||a.pointerId!==s.pointerId))return;let f;if(a.type==="touchmove"){if(f=[...a.changedTouches].filter(C=>C.identifier===s.touchId)[0],!f||f.identifier!==s.touchId)return}else f=a;if(!s.isTouched){s.startMoving&&s.isScrolling&&t.emit("touchMoveOpposite",a);return}const u=f.pageX,p=f.pageY;if(a.preventedByNestedSwiper){r.startX=u,r.startY=p;return}if(!t.allowTouchMove){a.target.matches(s.focusableElements)||(t.allowClick=!1),s.isTouched&&(Object.assign(r,{startX:u,startY:p,currentX:u,currentY:p}),s.touchStartTime=J());return}if(n.touchReleaseOnEdges&&!n.loop){if(t.isVertical()){if(p<r.startY&&t.translate<=t.maxTranslate()||p>r.startY&&t.translate>=t.minTranslate()){s.isTouched=!1,s.isMoved=!1;return}}else if(u<r.startX&&t.translate<=t.maxTranslate()||u>r.startX&&t.translate>=t.minTranslate())return}if(e.activeElement&&a.target===e.activeElement&&a.target.matches(s.focusableElements)){s.isMoved=!0,t.allowClick=!1;return}s.allowTouchCallbacks&&t.emit("touchMove",a),r.previousX=r.currentX,r.previousY=r.currentY,r.currentX=u,r.currentY=p;const h=r.currentX-r.startX,m=r.currentY-r.startY;if(t.params.threshold&&Math.sqrt(h**2+m**2)<t.params.threshold)return;if(typeof s.isScrolling>"u"){let C;t.isHorizontal()&&r.currentY===r.startY||t.isVertical()&&r.currentX===r.startX?s.isScrolling=!1:h*h+m*m>=25&&(C=Math.atan2(Math.abs(m),Math.abs(h))*180/Math.PI,s.isScrolling=t.isHorizontal()?C>n.touchAngle:90-C>n.touchAngle)}if(s.isScrolling&&t.emit("touchMoveOpposite",a),typeof s.startMoving>"u"&&(r.currentX!==r.startX||r.currentY!==r.startY)&&(s.startMoving=!0),s.isScrolling){s.isTouched=!1;return}if(!s.startMoving)return;t.allowClick=!1,!n.cssMode&&a.cancelable&&a.preventDefault(),n.touchMoveStopPropagation&&!n.nested&&a.stopPropagation();let y=t.isHorizontal()?h:m,g=t.isHorizontal()?r.currentX-r.previousX:r.currentY-r.previousY;n.oneWayMovement&&(y=Math.abs(y)*(o?1:-1),g=Math.abs(g)*(o?1:-1)),r.diff=y,y*=n.touchRatio,o&&(y=-y,g=-g);const S=t.touchesDirection;t.swipeDirection=y>0?"prev":"next",t.touchesDirection=g>0?"prev":"next";const w=t.params.loop&&!n.cssMode,d=t.touchesDirection==="next"&&t.allowSlideNext||t.touchesDirection==="prev"&&t.allowSlidePrev;if(!s.isMoved){if(w&&d&&t.loopFix({direction:t.swipeDirection}),s.startTranslate=t.getTranslate(),t.setTransition(0),t.animating){const C=new window.CustomEvent("transitionend",{bubbles:!0,cancelable:!0});t.wrapperEl.dispatchEvent(C)}s.allowMomentumBounce=!1,n.grabCursor&&(t.allowSlideNext===!0||t.allowSlidePrev===!0)&&t.setGrabCursor(!0),t.emit("sliderFirstMove",a)}let c;if(new Date().getTime(),s.isMoved&&s.allowThresholdMove&&S!==t.touchesDirection&&w&&d&&Math.abs(y)>=1){Object.assign(r,{startX:u,startY:p,currentX:u,currentY:p,startTranslate:s.currentTranslate}),s.loopSwapReset=!0,s.startTranslate=s.currentTranslate;return}t.emit("sliderMove",a),s.isMoved=!0,s.currentTranslate=y+s.startTranslate;let v=!0,T=n.resistanceRatio;if(n.touchReleaseOnEdges&&(T=0),y>0?(w&&d&&!c&&s.allowThresholdMove&&s.currentTranslate>(n.centeredSlides?t.minTranslate()-t.slidesSizesGrid[t.activeIndex+1]:t.minTranslate())&&t.loopFix({direction:"prev",setTranslate:!0,activeSlideIndex:0}),s.currentTranslate>t.minTranslate()&&(v=!1,n.resistance&&(s.currentTranslate=t.minTranslate()-1+(-t.minTranslate()+s.startTranslate+y)**T))):y<0&&(w&&d&&!c&&s.allowThresholdMove&&s.currentTranslate<(n.centeredSlides?t.maxTranslate()+t.slidesSizesGrid[t.slidesSizesGrid.length-1]:t.maxTranslate())&&t.loopFix({direction:"next",setTranslate:!0,activeSlideIndex:t.slides.length-(n.slidesPerView==="auto"?t.slidesPerViewDynamic():Math.ceil(parseFloat(n.slidesPerView,10)))}),s.currentTranslate<t.maxTranslate()&&(v=!1,n.resistance&&(s.currentTranslate=t.maxTranslate()+1-(t.maxTranslate()-s.startTranslate-y)**T))),v&&(a.preventedByNestedSwiper=!0),!t.allowSlideNext&&t.swipeDirection==="next"&&s.currentTranslate<s.startTranslate&&(s.currentTranslate=s.startTranslate),!t.allowSlidePrev&&t.swipeDirection==="prev"&&s.currentTranslate>s.startTranslate&&(s.currentTranslate=s.startTranslate),!t.allowSlidePrev&&!t.allowSlideNext&&(s.currentTranslate=s.startTranslate),n.threshold>0)if(Math.abs(y)>n.threshold||s.allowThresholdMove){if(!s.allowThresholdMove){s.allowThresholdMove=!0,r.startX=r.currentX,r.startY=r.currentY,s.currentTranslate=s.startTranslate,r.diff=t.isHorizontal()?r.currentX-r.startX:r.currentY-r.startY;return}}else{s.currentTranslate=s.startTranslate;return}!n.followFinger||n.cssMode||((n.freeMode&&n.freeMode.enabled&&t.freeMode||n.watchSlidesProgress)&&(t.updateActiveIndex(),t.updateSlidesClasses()),n.freeMode&&n.freeMode.enabled&&t.freeMode&&t.freeMode.onTouchMove(),t.updateProgress(s.currentTranslate),t.setTranslate(s.currentTranslate))}function Nt(i){const e=this,t=e.touchEventsData;let s=i;s.originalEvent&&(s=s.originalEvent);let n;if(s.type==="touchend"||s.type==="touchcancel"){if(n=[...s.changedTouches].filter(T=>T.identifier===t.touchId)[0],!n||n.identifier!==t.touchId)return}else{if(t.touchId!==null||s.pointerId!==t.pointerId)return;n=s}if(["pointercancel","pointerout","pointerleave","contextmenu"].includes(s.type)&&!(["pointercancel","contextmenu"].includes(s.type)&&(e.browser.isSafari||e.browser.isWebView)))return;t.pointerId=null,t.touchId=null;const{params:o,touches:l,rtlTranslate:a,slidesGrid:f,enabled:u}=e;if(!u||!o.simulateTouch&&s.pointerType==="mouse")return;if(t.allowTouchCallbacks&&e.emit("touchEnd",s),t.allowTouchCallbacks=!1,!t.isTouched){t.isMoved&&o.grabCursor&&e.setGrabCursor(!1),t.isMoved=!1,t.startMoving=!1;return}o.grabCursor&&t.isMoved&&t.isTouched&&(e.allowSlideNext===!0||e.allowSlidePrev===!0)&&e.setGrabCursor(!1);const p=J(),h=p-t.touchStartTime;if(e.allowClick){const T=s.path||s.composedPath&&s.composedPath();e.updateClickedSlide(T&&T[0]||s.target,T),e.emit("tap click",s),h<300&&p-t.lastClickTime<300&&e.emit("doubleTap doubleClick",s)}if(t.lastClickTime=J(),ce(()=>{e.destroyed||(e.allowClick=!0)}),!t.isTouched||!t.isMoved||!e.swipeDirection||l.diff===0&&!t.loopSwapReset||t.currentTranslate===t.startTranslate&&!t.loopSwapReset){t.isTouched=!1,t.isMoved=!1,t.startMoving=!1;return}t.isTouched=!1,t.isMoved=!1,t.startMoving=!1;let m;if(o.followFinger?m=a?e.translate:-e.translate:m=-t.currentTranslate,o.cssMode)return;if(o.freeMode&&o.freeMode.enabled){e.freeMode.onTouchEnd({currentPos:m});return}const y=m>=-e.maxTranslate()&&!e.params.loop;let g=0,S=e.slidesSizesGrid[0];for(let T=0;T<f.length;T+=T<o.slidesPerGroupSkip?1:o.slidesPerGroup){const C=T<o.slidesPerGroupSkip-1?1:o.slidesPerGroup;typeof f[T+C]<"u"?(y||m>=f[T]&&m<f[T+C])&&(g=T,S=f[T+C]-f[T]):(y||m>=f[T])&&(g=T,S=f[f.length-1]-f[f.length-2])}let w=null,d=null;o.rewind&&(e.isBeginning?d=o.virtual&&o.virtual.enabled&&e.virtual?e.virtual.slides.length-1:e.slides.length-1:e.isEnd&&(w=0));const c=(m-f[g])/S,v=g<o.slidesPerGroupSkip-1?1:o.slidesPerGroup;if(h>o.longSwipesMs){if(!o.longSwipes){e.slideTo(e.activeIndex);return}e.swipeDirection==="next"&&(c>=o.longSwipesRatio?e.slideTo(o.rewind&&e.isEnd?w:g+v):e.slideTo(g)),e.swipeDirection==="prev"&&(c>1-o.longSwipesRatio?e.slideTo(g+v):d!==null&&c<0&&Math.abs(c)>o.longSwipesRatio?e.slideTo(d):e.slideTo(g))}else{if(!o.shortSwipes){e.slideTo(e.activeIndex);return}e.navigation&&(s.target===e.navigation.nextEl||s.target===e.navigation.prevEl)?s.target===e.navigation.nextEl?e.slideTo(g+v):e.slideTo(g):(e.swipeDirection==="next"&&e.slideTo(w!==null?w:g+v),e.swipeDirection==="prev"&&e.slideTo(d!==null?d:g))}}function we(){const i=this,{params:e,el:t}=i;if(t&&t.offsetWidth===0)return;e.breakpoints&&i.setBreakpoint();const{allowSlideNext:s,allowSlidePrev:n,snapGrid:r}=i,o=i.virtual&&i.params.virtual.enabled;i.allowSlideNext=!0,i.allowSlidePrev=!0,i.updateSize(),i.updateSlides(),i.updateSlidesClasses();const l=o&&e.loop;(e.slidesPerView==="auto"||e.slidesPerView>1)&&i.isEnd&&!i.isBeginning&&!i.params.centeredSlides&&!l?i.slideTo(i.slides.length-1,0,!1,!0):i.params.loop&&!o?i.slideToLoop(i.realIndex,0,!1,!0):i.slideTo(i.activeIndex,0,!1,!0),i.autoplay&&i.autoplay.running&&i.autoplay.paused&&(clearTimeout(i.autoplay.resizeTimeout),i.autoplay.resizeTimeout=setTimeout(()=>{i.autoplay&&i.autoplay.running&&i.autoplay.paused&&i.autoplay.resume()},500)),i.allowSlidePrev=n,i.allowSlideNext=s,i.params.watchOverflow&&r!==i.snapGrid&&i.checkOverflow()}function _t(i){const e=this;e.enabled&&(e.allowClick||(e.params.preventClicks&&i.preventDefault(),e.params.preventClicksPropagation&&e.animating&&(i.stopPropagation(),i.stopImmediatePropagation())))}function Ht(){const i=this,{wrapperEl:e,rtlTranslate:t,enabled:s}=i;if(!s)return;i.previousTranslate=i.translate,i.isHorizontal()?i.translate=-e.scrollLeft:i.translate=-e.scrollTop,i.translate===0&&(i.translate=0),i.updateActiveIndex(),i.updateSlidesClasses();let n;const r=i.maxTranslate()-i.minTranslate();r===0?n=0:n=(i.translate-i.minTranslate())/r,n!==i.progress&&i.updateProgress(t?-i.translate:i.translate),i.emit("setTranslate",i.translate,!1)}function Rt(i){const e=this;K(e,i.target),!(e.params.cssMode||e.params.slidesPerView!=="auto"&&!e.params.autoHeight)&&e.update()}function Wt(){const i=this;i.documentTouchHandlerProceeded||(i.documentTouchHandlerProceeded=!0,i.params.touchReleaseOnEdges&&(i.el.style.touchAction="auto"))}const Le=(i,e)=>{const t=N(),{params:s,el:n,wrapperEl:r,device:o}=i,l=!!s.nested,a=e==="on"?"addEventListener":"removeEventListener",f=e;t[a]("touchstart",i.onDocumentTouchStart,{passive:!1,capture:l}),n[a]("touchstart",i.onTouchStart,{passive:!1}),n[a]("pointerdown",i.onTouchStart,{passive:!1}),t[a]("touchmove",i.onTouchMove,{passive:!1,capture:l}),t[a]("pointermove",i.onTouchMove,{passive:!1,capture:l}),t[a]("touchend",i.onTouchEnd,{passive:!0}),t[a]("pointerup",i.onTouchEnd,{passive:!0}),t[a]("pointercancel",i.onTouchEnd,{passive:!0}),t[a]("touchcancel",i.onTouchEnd,{passive:!0}),t[a]("pointerout",i.onTouchEnd,{passive:!0}),t[a]("pointerleave",i.onTouchEnd,{passive:!0}),t[a]("contextmenu",i.onTouchEnd,{passive:!0}),(s.preventClicks||s.preventClicksPropagation)&&n[a]("click",i.onClick,!0),s.cssMode&&r[a]("scroll",i.onScroll),s.updateOnWindowResize?i[f](o.ios||o.android?"resize orientationchange observerUpdate":"resize observerUpdate",we,!0):i[f]("observerUpdate",we,!0),n[a]("load",i.onLoad,{capture:!0})};function jt(){const i=this,{params:e}=i;i.onTouchStart=$t.bind(i),i.onTouchMove=Ft.bind(i),i.onTouchEnd=Nt.bind(i),i.onDocumentTouchStart=Wt.bind(i),e.cssMode&&(i.onScroll=Ht.bind(i)),i.onClick=_t.bind(i),i.onLoad=Rt.bind(i),Le(i,"on")}function qt(){Le(this,"off")}var Yt={attachEvents:jt,detachEvents:qt};const ye=(i,e)=>i.grid&&e.grid&&e.grid.rows>1;function Xt(){const i=this,{realIndex:e,initialized:t,params:s,el:n}=i,r=s.breakpoints;if(!r||r&&Object.keys(r).length===0)return;const o=i.getBreakpoint(r,i.params.breakpointsBase,i.el);if(!o||i.currentBreakpoint===o)return;const a=(o in r?r[o]:void 0)||i.originalParams,f=ye(i,s),u=ye(i,a),p=s.enabled;f&&!u?(n.classList.remove(`${s.containerModifierClass}grid`,`${s.containerModifierClass}grid-column`),i.emitContainerClasses()):!f&&u&&(n.classList.add(`${s.containerModifierClass}grid`),(a.grid.fill&&a.grid.fill==="column"||!a.grid.fill&&s.grid.fill==="column")&&n.classList.add(`${s.containerModifierClass}grid-column`),i.emitContainerClasses()),["navigation","pagination","scrollbar"].forEach(w=>{if(typeof a[w]>"u")return;const d=s[w]&&s[w].enabled,c=a[w]&&a[w].enabled;d&&!c&&i[w].disable(),!d&&c&&i[w].enable()});const h=a.direction&&a.direction!==s.direction,m=s.loop&&(a.slidesPerView!==s.slidesPerView||h),y=s.loop;h&&t&&i.changeDirection(),B(i.params,a);const g=i.params.enabled,S=i.params.loop;Object.assign(i,{allowTouchMove:i.params.allowTouchMove,allowSlideNext:i.params.allowSlideNext,allowSlidePrev:i.params.allowSlidePrev}),p&&!g?i.disable():!p&&g&&i.enable(),i.currentBreakpoint=o,i.emit("_beforeBreakpoint",a),t&&(m?(i.loopDestroy(),i.loopCreate(e),i.updateSlides()):!y&&S?(i.loopCreate(e),i.updateSlides()):y&&!S&&i.loopDestroy()),i.emit("breakpoint",a)}function Ut(i,e,t){if(e===void 0&&(e="window"),!i||e==="container"&&!t)return;let s=!1;const n=V(),r=e==="window"?n.innerHeight:t.clientHeight,o=Object.keys(i).map(l=>{if(typeof l=="string"&&l.indexOf("@")===0){const a=parseFloat(l.substr(1));return{value:r*a,point:l}}return{value:l,point:l}});o.sort((l,a)=>parseInt(l.value,10)-parseInt(a.value,10));for(let l=0;l<o.length;l+=1){const{point:a,value:f}=o[l];e==="window"?n.matchMedia(`(min-width: ${f}px)`).matches&&(s=a):f<=t.clientWidth&&(s=a)}return s||"max"}var Kt={setBreakpoint:Xt,getBreakpoint:Ut};function Jt(i,e){const t=[];return i.forEach(s=>{typeof s=="object"?Object.keys(s).forEach(n=>{s[n]&&t.push(e+n)}):typeof s=="string"&&t.push(e+s)}),t}function Qt(){const i=this,{classNames:e,params:t,rtl:s,el:n,device:r}=i,o=Jt(["initialized",t.direction,{"free-mode":i.params.freeMode&&t.freeMode.enabled},{autoheight:t.autoHeight},{rtl:s},{grid:t.grid&&t.grid.rows>1},{"grid-column":t.grid&&t.grid.rows>1&&t.grid.fill==="column"},{android:r.android},{ios:r.ios},{"css-mode":t.cssMode},{centered:t.cssMode&&t.centeredSlides},{"watch-progress":t.watchSlidesProgress}],t.containerModifierClass);e.push(...o),n.classList.add(...e),i.emitContainerClasses()}function Zt(){const i=this,{el:e,classNames:t}=i;e.classList.remove(...t),i.emitContainerClasses()}var ei={addClasses:Qt,removeClasses:Zt};function ti(){const i=this,{isLocked:e,params:t}=i,{slidesOffsetBefore:s}=t;if(s){const n=i.slides.length-1,r=i.slidesGrid[n]+i.slidesSizesGrid[n]+s*2;i.isLocked=i.size>r}else i.isLocked=i.snapGrid.length===1;t.allowSlideNext===!0&&(i.allowSlideNext=!i.isLocked),t.allowSlidePrev===!0&&(i.allowSlidePrev=!i.isLocked),e&&e!==i.isLocked&&(i.isEnd=!1),e!==i.isLocked&&i.emit(i.isLocked?"lock":"unlock")}var ii={checkOverflow:ti},Se={init:!0,direction:"horizontal",oneWayMovement:!1,touchEventsTarget:"wrapper",initialSlide:0,speed:300,cssMode:!1,updateOnWindowResize:!0,resizeObserver:!0,nested:!1,createElements:!1,eventsPrefix:"swiper",enabled:!0,focusableElements:"input, select, option, textarea, button, video, label",width:null,height:null,preventInteractionOnTransition:!1,userAgent:null,url:null,edgeSwipeDetection:!1,edgeSwipeThreshold:20,autoHeight:!1,setWrapperSize:!1,virtualTranslate:!1,effect:"slide",breakpoints:void 0,breakpointsBase:"window",spaceBetween:0,slidesPerView:1,slidesPerGroup:1,slidesPerGroupSkip:0,slidesPerGroupAuto:!1,centeredSlides:!1,centeredSlidesBounds:!1,slidesOffsetBefore:0,slidesOffsetAfter:0,normalizeSlideIndex:!0,centerInsufficientSlides:!1,watchOverflow:!0,roundLengths:!1,touchRatio:1,touchAngle:45,simulateTouch:!0,shortSwipes:!0,longSwipes:!0,longSwipesRatio:.5,longSwipesMs:300,followFinger:!0,allowTouchMove:!0,threshold:5,touchMoveStopPropagation:!1,touchStartPreventDefault:!0,touchStartForcePreventDefault:!1,touchReleaseOnEdges:!1,uniqueNavElements:!0,resistance:!0,resistanceRatio:.85,watchSlidesProgress:!1,grabCursor:!1,preventClicks:!0,preventClicksPropagation:!0,slideToClickedSlide:!1,loop:!1,loopAddBlankSlides:!0,loopAdditionalSlides:0,loopPreventsSliding:!0,rewind:!1,allowSlidePrev:!0,allowSlideNext:!0,swipeHandler:null,noSwiping:!0,noSwipingClass:"swiper-no-swiping",noSwipingSelector:null,passiveListeners:!0,maxBackfaceHiddenSlides:10,containerModifierClass:"swiper-",slideClass:"swiper-slide",slideBlankClass:"swiper-slide-blank",slideActiveClass:"swiper-slide-active",slideVisibleClass:"swiper-slide-visible",slideFullyVisibleClass:"swiper-slide-fully-visible",slideNextClass:"swiper-slide-next",slidePrevClass:"swiper-slide-prev",wrapperClass:"swiper-wrapper",lazyPreloaderClass:"swiper-lazy-preloader",lazyPreloadPrevNext:0,runCallbacksOnInit:!0,_emitClasses:!1};function si(i,e){return function(s){s===void 0&&(s={});const n=Object.keys(s)[0],r=s[n];if(typeof r!="object"||r===null){B(e,s);return}if(i[n]===!0&&(i[n]={enabled:!0}),n==="navigation"&&i[n]&&i[n].enabled&&!i[n].prevEl&&!i[n].nextEl&&(i[n].auto=!0),["pagination","scrollbar"].indexOf(n)>=0&&i[n]&&i[n].enabled&&!i[n].el&&(i[n].auto=!0),!(n in i&&"enabled"in r)){B(e,s);return}typeof i[n]=="object"&&!("enabled"in i[n])&&(i[n].enabled=!0),i[n]||(i[n]={enabled:!1}),B(e,s)}}const oe={eventsEmitter:et,update:ut,translate:vt,transition:Tt,slide:It,loop:kt,grabCursor:Bt,events:Yt,breakpoints:Kt,checkOverflow:ii,classes:ei},de={};class G{constructor(){let e,t;for(var s=arguments.length,n=new Array(s),r=0;r<s;r++)n[r]=arguments[r];n.length===1&&n[0].constructor&&Object.prototype.toString.call(n[0]).slice(8,-1)==="Object"?t=n[0]:[e,t]=n,t||(t={}),t=B({},t),e&&!t.el&&(t.el=e);const o=N();if(t.el&&typeof t.el=="string"&&o.querySelectorAll(t.el).length>1){const u=[];return o.querySelectorAll(t.el).forEach(p=>{const h=B({},t,{el:p});u.push(new G(h))}),u}const l=this;l.__swiper__=!0,l.support=Pe(),l.device=Ue({userAgent:t.userAgent}),l.browser=Je(),l.eventsListeners={},l.eventsAnyListeners=[],l.modules=[...l.__modules__],t.modules&&Array.isArray(t.modules)&&l.modules.push(...t.modules);const a={};l.modules.forEach(u=>{u({params:t,swiper:l,extendParams:si(t,a),on:l.on.bind(l),once:l.once.bind(l),off:l.off.bind(l),emit:l.emit.bind(l)})});const f=B({},Se,a);return l.params=B({},f,de,t),l.originalParams=B({},l.params),l.passedParams=B({},t),l.params&&l.params.on&&Object.keys(l.params.on).forEach(u=>{l.on(u,l.params.on[u])}),l.params&&l.params.onAny&&l.onAny(l.params.onAny),Object.assign(l,{enabled:l.params.enabled,el:e,classNames:[],slides:[],slidesGrid:[],snapGrid:[],slidesSizesGrid:[],isHorizontal(){return l.params.direction==="horizontal"},isVertical(){return l.params.direction==="vertical"},activeIndex:0,realIndex:0,isBeginning:!0,isEnd:!1,translate:0,previousTranslate:0,progress:0,velocity:0,animating:!1,cssOverflowAdjustment(){return Math.trunc(this.translate/2**23)*2**23},allowSlideNext:l.params.allowSlideNext,allowSlidePrev:l.params.allowSlidePrev,touchEventsData:{isTouched:void 0,isMoved:void 0,allowTouchCallbacks:void 0,touchStartTime:void 0,isScrolling:void 0,currentTranslate:void 0,startTranslate:void 0,allowThresholdMove:void 0,focusableElements:l.params.focusableElements,lastClickTime:0,clickTimeout:void 0,velocities:[],allowMomentumBounce:void 0,startMoving:void 0,pointerId:null,touchId:null},allowClick:!0,allowTouchMove:l.params.allowTouchMove,touches:{startX:0,startY:0,currentX:0,currentY:0,diff:0},imagesToLoad:[],imagesLoaded:0}),l.emit("_swiper"),l.params.init&&l.init(),l}getDirectionLabel(e){return this.isHorizontal()?e:{width:"height","margin-top":"margin-left","margin-bottom ":"margin-right","margin-left":"margin-top","margin-right":"margin-bottom","padding-left":"padding-top","padding-right":"padding-bottom",marginRight:"marginBottom"}[e]}getSlideIndex(e){const{slidesEl:t,params:s}=this,n=F(t,`.${s.slideClass}, swiper-slide`),r=ee(n[0]);return ee(e)-r}getSlideIndexByData(e){return this.getSlideIndex(this.slides.filter(t=>t.getAttribute("data-swiper-slide-index")*1===e)[0])}recalcSlides(){const e=this,{slidesEl:t,params:s}=e;e.slides=F(t,`.${s.slideClass}, swiper-slide`)}enable(){const e=this;e.enabled||(e.enabled=!0,e.params.grabCursor&&e.setGrabCursor(),e.emit("enable"))}disable(){const e=this;e.enabled&&(e.enabled=!1,e.params.grabCursor&&e.unsetGrabCursor(),e.emit("disable"))}setProgress(e,t){const s=this;e=Math.min(Math.max(e,0),1);const n=s.minTranslate(),o=(s.maxTranslate()-n)*e+n;s.translateTo(o,typeof t>"u"?0:t),s.updateActiveIndex(),s.updateSlidesClasses()}emitContainerClasses(){const e=this;if(!e.params._emitClasses||!e.el)return;const t=e.el.className.split(" ").filter(s=>s.indexOf("swiper")===0||s.indexOf(e.params.containerModifierClass)===0);e.emit("_containerClasses",t.join(" "))}getSlideClasses(e){const t=this;return t.destroyed?"":e.className.split(" ").filter(s=>s.indexOf("swiper-slide")===0||s.indexOf(t.params.slideClass)===0).join(" ")}emitSlidesClasses(){const e=this;if(!e.params._emitClasses||!e.el)return;const t=[];e.slides.forEach(s=>{const n=e.getSlideClasses(s);t.push({slideEl:s,classNames:n}),e.emit("_slideClass",s,n)}),e.emit("_slideClasses",t)}slidesPerViewDynamic(e,t){e===void 0&&(e="current"),t===void 0&&(t=!1);const s=this,{params:n,slides:r,slidesGrid:o,slidesSizesGrid:l,size:a,activeIndex:f}=s;let u=1;if(typeof n.slidesPerView=="number")return n.slidesPerView;if(n.centeredSlides){let p=r[f]?r[f].swiperSlideSize:0,h;for(let m=f+1;m<r.length;m+=1)r[m]&&!h&&(p+=r[m].swiperSlideSize,u+=1,p>a&&(h=!0));for(let m=f-1;m>=0;m-=1)r[m]&&!h&&(p+=r[m].swiperSlideSize,u+=1,p>a&&(h=!0))}else if(e==="current")for(let p=f+1;p<r.length;p+=1)(t?o[p]+l[p]-o[f]<a:o[p]-o[f]<a)&&(u+=1);else for(let p=f-1;p>=0;p-=1)o[f]-o[p]<a&&(u+=1);return u}update(){const e=this;if(!e||e.destroyed)return;const{snapGrid:t,params:s}=e;s.breakpoints&&e.setBreakpoint(),[...e.el.querySelectorAll('[loading="lazy"]')].forEach(o=>{o.complete&&K(e,o)}),e.updateSize(),e.updateSlides(),e.updateProgress(),e.updateSlidesClasses();function n(){const o=e.rtlTranslate?e.translate*-1:e.translate,l=Math.min(Math.max(o,e.maxTranslate()),e.minTranslate());e.setTranslate(l),e.updateActiveIndex(),e.updateSlidesClasses()}let r;if(s.freeMode&&s.freeMode.enabled&&!s.cssMode)n(),s.autoHeight&&e.updateAutoHeight();else{if((s.slidesPerView==="auto"||s.slidesPerView>1)&&e.isEnd&&!s.centeredSlides){const o=e.virtual&&s.virtual.enabled?e.virtual.slides:e.slides;r=e.slideTo(o.length-1,0,!1,!0)}else r=e.slideTo(e.activeIndex,0,!1,!0);r||n()}s.watchOverflow&&t!==e.snapGrid&&e.checkOverflow(),e.emit("update")}changeDirection(e,t){t===void 0&&(t=!0);const s=this,n=s.params.direction;return e||(e=n==="horizontal"?"vertical":"horizontal"),e===n||e!=="horizontal"&&e!=="vertical"||(s.el.classList.remove(`${s.params.containerModifierClass}${n}`),s.el.classList.add(`${s.params.containerModifierClass}${e}`),s.emitContainerClasses(),s.params.direction=e,s.slides.forEach(r=>{e==="vertical"?r.style.width="":r.style.height=""}),s.emit("changeDirection"),t&&s.update()),s}changeLanguageDirection(e){const t=this;t.rtl&&e==="rtl"||!t.rtl&&e==="ltr"||(t.rtl=e==="rtl",t.rtlTranslate=t.params.direction==="horizontal"&&t.rtl,t.rtl?(t.el.classList.add(`${t.params.containerModifierClass}rtl`),t.el.dir="rtl"):(t.el.classList.remove(`${t.params.containerModifierClass}rtl`),t.el.dir="ltr"),t.update())}mount(e){const t=this;if(t.mounted)return!0;let s=e||t.params.el;if(typeof s=="string"&&(s=document.querySelector(s)),!s)return!1;s.swiper=t,s.parentNode&&s.parentNode.host&&s.parentNode.host.nodeName==="SWIPER-CONTAINER"&&(t.isElement=!0);const n=()=>`.${(t.params.wrapperClass||"").trim().split(" ").join(".")}`;let o=s&&s.shadowRoot&&s.shadowRoot.querySelector?s.shadowRoot.querySelector(n()):F(s,n())[0];return!o&&t.params.createElements&&(o=Z("div",t.params.wrapperClass),s.append(o),F(s,`.${t.params.slideClass}`).forEach(l=>{o.append(l)})),Object.assign(t,{el:s,wrapperEl:o,slidesEl:t.isElement&&!s.parentNode.host.slideSlots?s.parentNode.host:o,hostEl:t.isElement?s.parentNode.host:s,mounted:!0,rtl:s.dir.toLowerCase()==="rtl"||_(s,"direction")==="rtl",rtlTranslate:t.params.direction==="horizontal"&&(s.dir.toLowerCase()==="rtl"||_(s,"direction")==="rtl"),wrongRTL:_(o,"display")==="-webkit-box"}),!0}init(e){const t=this;if(t.initialized||t.mount(e)===!1)return t;t.emit("beforeInit"),t.params.breakpoints&&t.setBreakpoint(),t.addClasses(),t.updateSize(),t.updateSlides(),t.params.watchOverflow&&t.checkOverflow(),t.params.grabCursor&&t.enabled&&t.setGrabCursor(),t.params.loop&&t.virtual&&t.params.virtual.enabled?t.slideTo(t.params.initialSlide+t.virtual.slidesBefore,0,t.params.runCallbacksOnInit,!1,!0):t.slideTo(t.params.initialSlide,0,t.params.runCallbacksOnInit,!1,!0),t.params.loop&&t.loopCreate(),t.attachEvents();const n=[...t.el.querySelectorAll('[loading="lazy"]')];return t.isElement&&n.push(...t.hostEl.querySelectorAll('[loading="lazy"]')),n.forEach(r=>{r.complete?K(t,r):r.addEventListener("load",o=>{K(t,o.target)})}),fe(t),t.initialized=!0,fe(t),t.emit("init"),t.emit("afterInit"),t}destroy(e,t){e===void 0&&(e=!0),t===void 0&&(t=!0);const s=this,{params:n,el:r,wrapperEl:o,slides:l}=s;return typeof s.params>"u"||s.destroyed||(s.emit("beforeDestroy"),s.initialized=!1,s.detachEvents(),n.loop&&s.loopDestroy(),t&&(s.removeClasses(),r.removeAttribute("style"),o.removeAttribute("style"),l&&l.length&&l.forEach(a=>{a.classList.remove(n.slideVisibleClass,n.slideFullyVisibleClass,n.slideActiveClass,n.slideNextClass,n.slidePrevClass),a.removeAttribute("style"),a.removeAttribute("data-swiper-slide-index")})),s.emit("destroy"),Object.keys(s.eventsListeners).forEach(a=>{s.off(a)}),e!==!1&&(s.el.swiper=null,_e(s)),s.destroyed=!0),null}static extendDefaults(e){B(de,e)}static get extendedDefaults(){return de}static get defaults(){return Se}static installModule(e){G.prototype.__modules__||(G.prototype.__modules__=[]);const t=G.prototype.__modules__;typeof e=="function"&&t.indexOf(e)<0&&t.push(e)}static use(e){return Array.isArray(e)?(e.forEach(t=>G.installModule(t)),G):(G.installModule(e),G)}}Object.keys(oe).forEach(i=>{Object.keys(oe[i]).forEach(e=>{G.prototype[e]=oe[i][e]})});G.use([Qe,Ze]);function Ie(i,e,t,s){return i.params.createElements&&Object.keys(s).forEach(n=>{if(!t[n]&&t.auto===!0){let r=F(i.el,`.${s[n]}`)[0];r||(r=Z("div",s[n]),r.className=s[n],i.el.append(r)),t[n]=r,e[n]=r}}),t}function te(i){let{swiper:e,extendParams:t,on:s,emit:n}=i;t({navigation:{nextEl:null,prevEl:null,hideOnClick:!1,disabledClass:"swiper-button-disabled",hiddenClass:"swiper-button-hidden",lockClass:"swiper-button-lock",navigationDisabledClass:"swiper-navigation-disabled"}}),e.navigation={nextEl:null,prevEl:null};const r=g=>(Array.isArray(g)?g:[g]).filter(S=>!!S);function o(g){let S;return g&&typeof g=="string"&&e.isElement&&(S=e.el.querySelector(g),S)?S:(g&&(typeof g=="string"&&(S=[...document.querySelectorAll(g)]),e.params.uniqueNavElements&&typeof g=="string"&&S.length>1&&e.el.querySelectorAll(g).length===1&&(S=e.el.querySelector(g))),g&&!S?g:S)}function l(g,S){const w=e.params.navigation;g=r(g),g.forEach(d=>{d&&(d.classList[S?"add":"remove"](...w.disabledClass.split(" ")),d.tagName==="BUTTON"&&(d.disabled=S),e.params.watchOverflow&&e.enabled&&d.classList[e.isLocked?"add":"remove"](w.lockClass))})}function a(){const{nextEl:g,prevEl:S}=e.navigation;if(e.params.loop){l(S,!1),l(g,!1);return}l(S,e.isBeginning&&!e.params.rewind),l(g,e.isEnd&&!e.params.rewind)}function f(g){g.preventDefault(),!(e.isBeginning&&!e.params.loop&&!e.params.rewind)&&(e.slidePrev(),n("navigationPrev"))}function u(g){g.preventDefault(),!(e.isEnd&&!e.params.loop&&!e.params.rewind)&&(e.slideNext(),n("navigationNext"))}function p(){const g=e.params.navigation;if(e.params.navigation=Ie(e,e.originalParams.navigation,e.params.navigation,{nextEl:"swiper-button-next",prevEl:"swiper-button-prev"}),!(g.nextEl||g.prevEl))return;let S=o(g.nextEl),w=o(g.prevEl);Object.assign(e.navigation,{nextEl:S,prevEl:w}),S=r(S),w=r(w);const d=(c,v)=>{c&&c.addEventListener("click",v==="next"?u:f),!e.enabled&&c&&c.classList.add(...g.lockClass.split(" "))};S.forEach(c=>d(c,"next")),w.forEach(c=>d(c,"prev"))}function h(){let{nextEl:g,prevEl:S}=e.navigation;g=r(g),S=r(S);const w=(d,c)=>{d.removeEventListener("click",c==="next"?u:f),d.classList.remove(...e.params.navigation.disabledClass.split(" "))};g.forEach(d=>w(d,"next")),S.forEach(d=>w(d,"prev"))}s("init",()=>{e.params.navigation.enabled===!1?y():(p(),a())}),s("toEdge fromEdge lock unlock",()=>{a()}),s("destroy",()=>{h()}),s("enable disable",()=>{let{nextEl:g,prevEl:S}=e.navigation;if(g=r(g),S=r(S),e.enabled){a();return}[...g,...S].filter(w=>!!w).forEach(w=>w.classList.add(e.params.navigation.lockClass))}),s("click",(g,S)=>{let{nextEl:w,prevEl:d}=e.navigation;w=r(w),d=r(d);const c=S.target;if(e.params.navigation.hideOnClick&&!d.includes(c)&&!w.includes(c)){if(e.pagination&&e.params.pagination&&e.params.pagination.clickable&&(e.pagination.el===c||e.pagination.el.contains(c)))return;let v;w.length?v=w[0].classList.contains(e.params.navigation.hiddenClass):d.length&&(v=d[0].classList.contains(e.params.navigation.hiddenClass)),n(v===!0?"navigationShow":"navigationHide"),[...w,...d].filter(T=>!!T).forEach(T=>T.classList.toggle(e.params.navigation.hiddenClass))}});const m=()=>{e.el.classList.remove(...e.params.navigation.navigationDisabledClass.split(" ")),p(),a()},y=()=>{e.el.classList.add(...e.params.navigation.navigationDisabledClass.split(" ")),h()};Object.assign(e.navigation,{enable:m,disable:y,update:a,init:p,destroy:h})}function W(i){return i===void 0&&(i=""),`.${i.trim().replace(/([\.:!+\/])/g,"\\$1").replace(/ /g,".")}`}function ie(i){let{swiper:e,extendParams:t,on:s,emit:n}=i;const r="swiper-pagination";t({pagination:{el:null,bulletElement:"span",clickable:!1,hideOnClick:!1,renderBullet:null,renderProgressbar:null,renderFraction:null,renderCustom:null,progressbarOpposite:!1,type:"bullets",dynamicBullets:!1,dynamicMainBullets:1,formatFractionCurrent:d=>d,formatFractionTotal:d=>d,bulletClass:`${r}-bullet`,bulletActiveClass:`${r}-bullet-active`,modifierClass:`${r}-`,currentClass:`${r}-current`,totalClass:`${r}-total`,hiddenClass:`${r}-hidden`,progressbarFillClass:`${r}-progressbar-fill`,progressbarOppositeClass:`${r}-progressbar-opposite`,clickableClass:`${r}-clickable`,lockClass:`${r}-lock`,horizontalClass:`${r}-horizontal`,verticalClass:`${r}-vertical`,paginationDisabledClass:`${r}-disabled`}}),e.pagination={el:null,bullets:[]};let o,l=0;const a=d=>(Array.isArray(d)?d:[d]).filter(c=>!!c);function f(){return!e.params.pagination.el||!e.pagination.el||Array.isArray(e.pagination.el)&&e.pagination.el.length===0}function u(d,c){const{bulletActiveClass:v}=e.params.pagination;d&&(d=d[`${c==="prev"?"previous":"next"}ElementSibling`],d&&(d.classList.add(`${v}-${c}`),d=d[`${c==="prev"?"previous":"next"}ElementSibling`],d&&d.classList.add(`${v}-${c}-${c}`)))}function p(d){const c=d.target.closest(W(e.params.pagination.bulletClass));if(!c)return;d.preventDefault();const v=ee(c)*e.params.slidesPerGroup;if(e.params.loop){if(e.realIndex===v)return;e.slideToLoop(v)}else e.slideTo(v)}function h(){const d=e.rtl,c=e.params.pagination;if(f())return;let v=e.pagination.el;v=a(v);let T,C;const z=e.virtual&&e.params.virtual.enabled?e.virtual.slides.length:e.slides.length,A=e.params.loop?Math.ceil(z/e.params.slidesPerGroup):e.snapGrid.length;if(e.params.loop?(C=e.previousRealIndex||0,T=e.params.slidesPerGroup>1?Math.floor(e.realIndex/e.params.slidesPerGroup):e.realIndex):typeof e.snapIndex<"u"?(T=e.snapIndex,C=e.previousSnapIndex):(C=e.previousIndex||0,T=e.activeIndex||0),c.type==="bullets"&&e.pagination.bullets&&e.pagination.bullets.length>0){const L=e.pagination.bullets;let b,E,P;if(c.dynamicBullets&&(o=ue(L[0],e.isHorizontal()?"width":"height",!0),v.forEach(x=>{x.style[e.isHorizontal()?"width":"height"]=`${o*(c.dynamicMainBullets+4)}px`}),c.dynamicMainBullets>1&&C!==void 0&&(l+=T-(C||0),l>c.dynamicMainBullets-1?l=c.dynamicMainBullets-1:l<0&&(l=0)),b=Math.max(T-l,0),E=b+(Math.min(L.length,c.dynamicMainBullets)-1),P=(E+b)/2),L.forEach(x=>{const M=[...["","-next","-next-next","-prev","-prev-prev","-main"].map(I=>`${c.bulletActiveClass}${I}`)].map(I=>typeof I=="string"&&I.includes(" ")?I.split(" "):I).flat();x.classList.remove(...M)}),v.length>1)L.forEach(x=>{const M=ee(x);M===T?x.classList.add(...c.bulletActiveClass.split(" ")):e.isElement&&x.setAttribute("part","bullet"),c.dynamicBullets&&(M>=b&&M<=E&&x.classList.add(...`${c.bulletActiveClass}-main`.split(" ")),M===b&&u(x,"prev"),M===E&&u(x,"next"))});else{const x=L[T];if(x&&x.classList.add(...c.bulletActiveClass.split(" ")),e.isElement&&L.forEach((M,I)=>{M.setAttribute("part",I===T?"bullet-active":"bullet")}),c.dynamicBullets){const M=L[b],I=L[E];for(let k=b;k<=E;k+=1)L[k]&&L[k].classList.add(...`${c.bulletActiveClass}-main`.split(" "));u(M,"prev"),u(I,"next")}}if(c.dynamicBullets){const x=Math.min(L.length,c.dynamicMainBullets+4),M=(o*x-o)/2-P*o,I=d?"right":"left";L.forEach(k=>{k.style[e.isHorizontal()?I:"top"]=`${M}px`})}}v.forEach((L,b)=>{if(c.type==="fraction"&&(L.querySelectorAll(W(c.currentClass)).forEach(E=>{E.textContent=c.formatFractionCurrent(T+1)}),L.querySelectorAll(W(c.totalClass)).forEach(E=>{E.textContent=c.formatFractionTotal(A)})),c.type==="progressbar"){let E;c.progressbarOpposite?E=e.isHorizontal()?"vertical":"horizontal":E=e.isHorizontal()?"horizontal":"vertical";const P=(T+1)/A;let x=1,M=1;E==="horizontal"?x=P:M=P,L.querySelectorAll(W(c.progressbarFillClass)).forEach(I=>{I.style.transform=`translate3d(0,0,0) scaleX(${x}) scaleY(${M})`,I.style.transitionDuration=`${e.params.speed}ms`})}c.type==="custom"&&c.renderCustom?(L.innerHTML=c.renderCustom(e,T+1,A),b===0&&n("paginationRender",L)):(b===0&&n("paginationRender",L),n("paginationUpdate",L)),e.params.watchOverflow&&e.enabled&&L.classList[e.isLocked?"add":"remove"](c.lockClass)})}function m(){const d=e.params.pagination;if(f())return;const c=e.virtual&&e.params.virtual.enabled?e.virtual.slides.length:e.grid&&e.params.grid.rows>1?e.slides.length/Math.ceil(e.params.grid.rows):e.slides.length;let v=e.pagination.el;v=a(v);let T="";if(d.type==="bullets"){let C=e.params.loop?Math.ceil(c/e.params.slidesPerGroup):e.snapGrid.length;e.params.freeMode&&e.params.freeMode.enabled&&C>c&&(C=c);for(let z=0;z<C;z+=1)d.renderBullet?T+=d.renderBullet.call(e,z,d.bulletClass):T+=`<${d.bulletElement} ${e.isElement?'part="bullet"':""} class="${d.bulletClass}"></${d.bulletElement}>`}d.type==="fraction"&&(d.renderFraction?T=d.renderFraction.call(e,d.currentClass,d.totalClass):T=`<span class="${d.currentClass}"></span> / <span class="${d.totalClass}"></span>`),d.type==="progressbar"&&(d.renderProgressbar?T=d.renderProgressbar.call(e,d.progressbarFillClass):T=`<span class="${d.progressbarFillClass}"></span>`),e.pagination.bullets=[],v.forEach(C=>{d.type!=="custom"&&(C.innerHTML=T||""),d.type==="bullets"&&e.pagination.bullets.push(...C.querySelectorAll(W(d.bulletClass)))}),d.type!=="custom"&&n("paginationRender",v[0])}function y(){e.params.pagination=Ie(e,e.originalParams.pagination,e.params.pagination,{el:"swiper-pagination"});const d=e.params.pagination;if(!d.el)return;let c;typeof d.el=="string"&&e.isElement&&(c=e.el.querySelector(d.el)),!c&&typeof d.el=="string"&&(c=[...document.querySelectorAll(d.el)]),c||(c=d.el),!(!c||c.length===0)&&(e.params.uniqueNavElements&&typeof d.el=="string"&&Array.isArray(c)&&c.length>1&&(c=[...e.el.querySelectorAll(d.el)],c.length>1&&(c=c.filter(v=>Ce(v,".swiper")[0]===e.el)[0])),Array.isArray(c)&&c.length===1&&(c=c[0]),Object.assign(e.pagination,{el:c}),c=a(c),c.forEach(v=>{d.type==="bullets"&&d.clickable&&v.classList.add(...(d.clickableClass||"").split(" ")),v.classList.add(d.modifierClass+d.type),v.classList.add(e.isHorizontal()?d.horizontalClass:d.verticalClass),d.type==="bullets"&&d.dynamicBullets&&(v.classList.add(`${d.modifierClass}${d.type}-dynamic`),l=0,d.dynamicMainBullets<1&&(d.dynamicMainBullets=1)),d.type==="progressbar"&&d.progressbarOpposite&&v.classList.add(d.progressbarOppositeClass),d.clickable&&v.addEventListener("click",p),e.enabled||v.classList.add(d.lockClass)}))}function g(){const d=e.params.pagination;if(f())return;let c=e.pagination.el;c&&(c=a(c),c.forEach(v=>{v.classList.remove(d.hiddenClass),v.classList.remove(d.modifierClass+d.type),v.classList.remove(e.isHorizontal()?d.horizontalClass:d.verticalClass),d.clickable&&(v.classList.remove(...(d.clickableClass||"").split(" ")),v.removeEventListener("click",p))})),e.pagination.bullets&&e.pagination.bullets.forEach(v=>v.classList.remove(...d.bulletActiveClass.split(" ")))}s("changeDirection",()=>{if(!e.pagination||!e.pagination.el)return;const d=e.params.pagination;let{el:c}=e.pagination;c=a(c),c.forEach(v=>{v.classList.remove(d.horizontalClass,d.verticalClass),v.classList.add(e.isHorizontal()?d.horizontalClass:d.verticalClass)})}),s("init",()=>{e.params.pagination.enabled===!1?w():(y(),m(),h())}),s("activeIndexChange",()=>{typeof e.snapIndex>"u"&&h()}),s("snapIndexChange",()=>{h()}),s("snapGridLengthChange",()=>{m(),h()}),s("destroy",()=>{g()}),s("enable disable",()=>{let{el:d}=e.pagination;d&&(d=a(d),d.forEach(c=>c.classList[e.enabled?"remove":"add"](e.params.pagination.lockClass)))}),s("lock unlock",()=>{h()}),s("click",(d,c)=>{const v=c.target,T=a(e.pagination.el);if(e.params.pagination.el&&e.params.pagination.hideOnClick&&T&&T.length>0&&!v.classList.contains(e.params.pagination.bulletClass)){if(e.navigation&&(e.navigation.nextEl&&v===e.navigation.nextEl||e.navigation.prevEl&&v===e.navigation.prevEl))return;const C=T[0].classList.contains(e.params.pagination.hiddenClass);n(C===!0?"paginationShow":"paginationHide"),T.forEach(z=>z.classList.toggle(e.params.pagination.hiddenClass))}});const S=()=>{e.el.classList.remove(e.params.pagination.paginationDisabledClass);let{el:d}=e.pagination;d&&(d=a(d),d.forEach(c=>c.classList.remove(e.params.pagination.paginationDisabledClass))),y(),m(),h()},w=()=>{e.el.classList.add(e.params.pagination.paginationDisabledClass);let{el:d}=e.pagination;d&&(d=a(d),d.forEach(c=>c.classList.add(e.params.pagination.paginationDisabledClass))),g()};Object.assign(e.pagination,{enable:S,disable:w,render:m,update:h,init:y,destroy:g})}function me(i){let{swiper:e,extendParams:t,on:s,emit:n,params:r}=i;e.autoplay={running:!1,paused:!1,timeLeft:0},t({autoplay:{enabled:!1,delay:3e3,waitForTransition:!0,disableOnInteraction:!1,stopOnLastSlide:!1,reverseDirection:!1,pauseOnMouseEnter:!1}});let o,l,a=r&&r.autoplay?r.autoplay.delay:3e3,f=r&&r.autoplay?r.autoplay.delay:3e3,u,p=new Date().getTime(),h,m,y,g,S,w,d;function c(O){!e||e.destroyed||!e.wrapperEl||O.target===e.wrapperEl&&(e.wrapperEl.removeEventListener("transitionend",c),!d&&b())}const v=()=>{if(e.destroyed||!e.autoplay.running)return;e.autoplay.paused?h=!0:h&&(f=u,h=!1);const O=e.autoplay.paused?u:p+f-new Date().getTime();e.autoplay.timeLeft=O,n("autoplayTimeLeft",O,O/a),l=requestAnimationFrame(()=>{v()})},T=()=>{let O;return e.virtual&&e.params.virtual.enabled?O=e.slides.filter(D=>D.classList.contains("swiper-slide-active"))[0]:O=e.slides[e.activeIndex],O?parseInt(O.getAttribute("data-swiper-autoplay"),10):void 0},C=O=>{if(e.destroyed||!e.autoplay.running)return;cancelAnimationFrame(l),v();let $=typeof O>"u"?e.params.autoplay.delay:O;a=e.params.autoplay.delay,f=e.params.autoplay.delay;const D=T();!Number.isNaN(D)&&D>0&&typeof O>"u"&&($=D,a=D,f=D),u=$;const H=e.params.speed,Y=()=>{!e||e.destroyed||(e.params.autoplay.reverseDirection?!e.isBeginning||e.params.loop||e.params.rewind?(e.slidePrev(H,!0,!0),n("autoplay")):e.params.autoplay.stopOnLastSlide||(e.slideTo(e.slides.length-1,H,!0,!0),n("autoplay")):!e.isEnd||e.params.loop||e.params.rewind?(e.slideNext(H,!0,!0),n("autoplay")):e.params.autoplay.stopOnLastSlide||(e.slideTo(0,H,!0,!0),n("autoplay")),e.params.cssMode&&(p=new Date().getTime(),requestAnimationFrame(()=>{C()})))};return $>0?(clearTimeout(o),o=setTimeout(()=>{Y()},$)):requestAnimationFrame(()=>{Y()}),$},z=()=>{p=new Date().getTime(),e.autoplay.running=!0,C(),n("autoplayStart")},A=()=>{e.autoplay.running=!1,clearTimeout(o),cancelAnimationFrame(l),n("autoplayStop")},L=(O,$)=>{if(e.destroyed||!e.autoplay.running)return;clearTimeout(o),O||(w=!0);const D=()=>{n("autoplayPause"),e.params.autoplay.waitForTransition?e.wrapperEl.addEventListener("transitionend",c):b()};if(e.autoplay.paused=!0,$){S&&(u=e.params.autoplay.delay),S=!1,D();return}u=(u||e.params.autoplay.delay)-(new Date().getTime()-p),!(e.isEnd&&u<0&&!e.params.loop)&&(u<0&&(u=0),D())},b=()=>{e.isEnd&&u<0&&!e.params.loop||e.destroyed||!e.autoplay.running||(p=new Date().getTime(),w?(w=!1,C(u)):C(),e.autoplay.paused=!1,n("autoplayResume"))},E=()=>{if(e.destroyed||!e.autoplay.running)return;const O=N();O.visibilityState==="hidden"&&(w=!0,L(!0)),O.visibilityState==="visible"&&b()},P=O=>{O.pointerType==="mouse"&&(w=!0,d=!0,!(e.animating||e.autoplay.paused)&&L(!0))},x=O=>{O.pointerType==="mouse"&&(d=!1,e.autoplay.paused&&b())},M=()=>{e.params.autoplay.pauseOnMouseEnter&&(e.el.addEventListener("pointerenter",P),e.el.addEventListener("pointerleave",x))},I=()=>{e.el.removeEventListener("pointerenter",P),e.el.removeEventListener("pointerleave",x)},k=()=>{N().addEventListener("visibilitychange",E)},se=()=>{N().removeEventListener("visibilitychange",E)};s("init",()=>{e.params.autoplay.enabled&&(M(),k(),z())}),s("destroy",()=>{I(),se(),e.autoplay.running&&A()}),s("_freeModeStaticRelease",()=>{(y||w)&&b()}),s("_freeModeNoMomentumRelease",()=>{e.params.autoplay.disableOnInteraction?A():L(!0,!0)}),s("beforeTransitionStart",(O,$,D)=>{e.destroyed||!e.autoplay.running||(D||!e.params.autoplay.disableOnInteraction?L(!0,!0):A())}),s("sliderFirstMove",()=>{if(!(e.destroyed||!e.autoplay.running)){if(e.params.autoplay.disableOnInteraction){A();return}m=!0,y=!1,w=!1,g=setTimeout(()=>{w=!0,y=!0,L(!0)},200)}}),s("touchEnd",()=>{if(!(e.destroyed||!e.autoplay.running||!m)){if(clearTimeout(g),clearTimeout(o),e.params.autoplay.disableOnInteraction){y=!1,m=!1;return}y&&e.params.cssMode&&b(),y=!1,m=!1}}),s("slideChange",()=>{e.destroyed||!e.autoplay.running||(S=!0)}),Object.assign(e.autoplay,{start:z,stop:A,pause:L,resume:b})}$e("칼리 | 마켓칼리");De();localStorage.getItem("view")||Be("view",Ge);const ri=await Ve("view");let ni=ri.id,Te=[];async function ai(){const i=await he.collection("products").getFullList();i.forEach(e=>{Oe(".recommend > .swiper-wrapper",e)}),i.forEach(e=>{ze(".sale > .swiper-wrapper",e)});for(let e of ni){const t=await he.collection("products").getOne(e);Te.push(t)}Te.forEach(e=>{ke(".recent > .swiper-wrapper",e)})}ai();new G(".swiper.banner",{modules:[te,ie,me],loop:!0,navigation:{nextEl:".swiper-button-next.banner",prevEl:".swiper-button-prev.banner"},autoplay:{delay:2500,disableOnInteraction:!1},pagination:{el:".swiper-pagination.banner",type:"fraction"}});new G(".swiper.recommend",{modules:[te,ie],loop:!1,slidesPerView:4,slidesPerGroup:4,spaceBetween:30,centeredSlides:!1,navigation:{nextEl:".swiper-button-next.recommend",prevEl:".swiper-button-prev.recommend"}});new G(".swiper.sale",{modules:[te,ie,me],loop:!0,slidesPerView:4,spaceBetween:30,centeredSlides:!0,autoplay:{delay:1e3,disableOnInteraction:!1},navigation:{nextEl:".swiper-button-next.sale",prevEl:".swiper-button-prev.sale"}});new G(".swiper.recent",{modules:[te,ie,me],direction:"vertical",slidesPerView:3,spaceBetween:15,centeredSlides:!1,navigation:{nextEl:".swiper-button-next.recent",prevEl:".swiper-button-prev.recent"}});const j=be(".swiper-button-prev.recommend"),q=be(".swiper-button-next.recommend");function Ae(){q.classList.contains("swiper-button-disabled")?R(q,"display","none"):R(q,"display","flex"),j.classList.contains("swiper-button-disabled")?R(j,"display","none"):R(j,"display","flex")}(function(){R(q,"display","flex"),R(j,"display","none")})();j.addEventListener("click",Ae);q.addEventListener("click",Ae);
