+++ docs/assets/register-61AnPVx2.js
@@ -1,0 +1,2 @@
+import"./style-v06b81ne.js";import{g as l,p as _}from"./header-tMjJ4wC4.js";/* empty css               */import{s as P}from"./setDocumentTitle-_OwlDZqT.js";P("칼리 | 회원가입");const a=l("html"),m=l(".user_id_input"),p=l(".user_password_input"),S=l(".user_password_check_input"),w=l(".user_name_input"),u=l(".user_email_input"),v=l(".user_phone_input"),L=l(".duplicate_id_btn"),g=l(".duplicate_alert_container"),y=l(".duplicate_alert_text"),k=l(".duplicate_email_btn"),h=l(".duplicate_email_alert_container"),b=l(".duplicate_email_alert_text"),$=l(".duplicate_alert_btn"),B=l(".duplicate_email_alert_btn"),n=l(".check_number_btn"),E=l(".error_message"),r=l("#user_password_error"),A=l("#user_password_check_error"),x=l("#user_email_error"),f=l("#agree_all");function d(e,t){e.target.value&&t.classList.add("input_active")}function C(e){return!!/^(?=.*[A-Za-z])[A-Za-z\d]{6,16}$/.test(e)}function V(e){return!!/^(?=.*[!@#$%^&*(),.?":{}|<>]).{6,16}$/.test(e)}function I(e){return!!/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(e)}async function j(){const e=await _.collection("users").getFullList();return console.log(e),e.filter(i=>m.value===i.username).length!==0}async function T(){C(m.value)?await j()?(y.textContent="이미 등록된 아이디 입니다",g.style.visibility="visible",a.style.overflow="hidden"):(L.disabled=!0,L.classList.add("disable_btn"),y.textContent="사용 할 수 있는 아이디 입니다",g.style.visibility="visible",a.style.overflow="hidden"):(y.textContent="6자 이상 16자 이하의 영문 혹은 영문과 숫자를 조합",g.style.visibility="visible",a.style.overflow="hidden")}async function z(){const e={username:"",password:"123456!",passwordConfirm:"123456!",email:u.value,emailVisibility:!0};I(u.value)?_.collection("users").create(e).then(()=>{k.disabled=!0,k.classList.add("disable_btn"),b.textContent="사용 가능한 이메일 입니다.",h.style.visibility="visible",a.style.overflow="hidden",D(),console.log("성공!")}).catch(t=>{b.textContent="이미 등록된 이메일 입니다.",h.style.visibility="visible",a.style.overflow="hidden",console.log("이메일 만들기 실패",t)}):(b.textContent="이메일 형식으로 입력해 주세요.",h.style.visibility="visible",a.style.overflow="hidden")}async function D(){let t=(await _.collection("users").getFullList()).filter(i=>u.value===i.email);t.length!==0&&await _.collection("users").delete(t[0].id)}async function Z(){n.disabled||(document.querySelector(".duplicate_number_alert_text").textContent="믿음으로 인증되었습니다",document.querySelector(".duplicate_number_alert_container").style.visibility="visible",n.classList.add("disable_btn"),a.style.overflow="hidden",n.disabled=!0)}function F(){g.style.visibility="hidden",a.style.overflow="scroll"}function K(){h.style.visibility="hidden",a.style.overflow="scroll"}function U(){document.querySelector(".duplicate_number_alert_container").style.visibility="hidden",a.style.overflow="scroll"}m.addEventListener("input",e=>{d(e,m),C(e.target.value)?E.classList.remove("is_invalid"):E.classList.add("is_invalid")});p.addEventListener("input",e=>{d(e,p);let t=V(e.target.value);e.target.value&&e.target.value.length<6?(r.classList.add("is_invalid"),r.textContent="최소 6자 이상 입력"):e.target.value.length>16?(r.classList.add("is_invalid"),r.textContent="최대 16자 까지 입력"):t?r.classList.remove("is_invalid"):r.textContent="특수문자 포함 최소 6자 - 최대 16자"});S.addEventListener("input",e=>{d(e,S),e.target.value===p.value?A.classList.remove("is_invalid"):A.classList.add("is_invalid")});w.addEventListener("input",e=>d(e,w));u.addEventListener("input",e=>{d(e,u),I(e.target.value)?x.classList.remove("is_invalid"):x.classList.add("is_invalid")});v.addEventListener("input",e=>{d(e,v);let t=e.target.value;isNaN(t)?v.value="":(v.value=t,t?(n.classList.remove("disable_btn"),n.disabled=!1):(n.classList.add("disable_btn"),n.disabled=!0))});n.addEventListener("click",Z);document.querySelector(".duplicate_number_alert_btn").addEventListener("click",U);L.addEventListener("click",T);$.addEventListener("click",F);k.addEventListener("click",z);B.addEventListener("click",K);function q(e){for(let t=1;t<5;t++)document.querySelector(`#check_${t}`).checked=e}f.addEventListener("click",function(){f.checked?q(!0):q(!1)});function M(){new daum.Postcode({oncomplete:function(e){e.userSelectedType}}).open({popupKey:"popup1",popupTitle:"컬리 - 마켓컬리/뷰티컬리",left:window.screen.width/2-530/2,top:window.screen.height/2-630/2})}const R=l(".user_address_btn");R.addEventListener("click",M);document.querySelector(".join_btn").addEventListener("click",function(e){e.preventDefault();const t={username:m.value,password:p.value,passwordConfirm:p.value,name:w.value,email:u.value,emailVisibility:!0,phone:v.value,gender:document.querySelectorAll(".radio_wrapper")[0].value??"none",birth:`${document.querySelector("#year").value}${document.querySelector("#month").value}${document.querySelector("#day").value}`};_.collection("users").create(t).then(()=>{console.log("회원가입 완료!"),location.href="/Karly/src/pages/login/"}).catch(i=>{console.log("회원가입 실패",i)})});document.querySelectorAll(".radio_wrapper")[0].addEventListener("click",function(e){e.target.tagName==="INPUT"&&(document.querySelectorAll(".radio_wrapper")[0].value=e.target.value)});document.querySelector("#year").addEventListener("input",function(e){if(e.target.value.length>4){e.target.value=e.target.value.toString(),e.target.value=e.target.value.slice(0,4);let t=Number(e.target.value);t>=1900&&t<=2024&&(document.querySelector("#year").textContent=e.target.value)}});document.querySelector("#month").addEventListener("input",function(e){if(e.target.value.length>2){e.target.value=e.target.value.toString(),e.target.value=e.target.value.slice(0,2);let t=Number(e.target.value);t>=1&&t<=12&&(document.querySelector("#month").textContent=e.target.value)}});document.querySelector("#day").addEventListener("input",function(e){if(e.target.value.length>2){e.target.value=e.target.value.toString(),e.target.value=e.target.value.slice(0,2);let t=Number(e.target.value);t>=1&&t<=31&&(document.querySelector("#day").textContent=e.target.value)}});let G=!1,s=!1,o=!1,c=!1;document.querySelector(".agree_info_wrapper").addEventListener("change",function(e){e.target.name==="agree_all"?e.target.checked?(s=!0,o=!0,c=!0):(s=!1,o=!1,c=!1):e.target.name==="check_1"?e.target.checked?s=!0:(s=!1,f.checked=!1):e.target.name==="check_2"?e.target.checked?o=!0:(o=!1,f.checked=!1):e.target.name==="check_4"&&(e.target.checked?c=!0:(c=!1,f.checked=!1)),console.log("체크1",s),console.log("체크2",o),console.log("체크4",c),console.log(`
+`),H(s,o,c,G)});function H(e,t,i,N){e&&t&&i||N?(document.querySelector(".join_btn").classList.remove("disable_btn"),document.querySelector(".join_btn").disabled=!1):(document.querySelector(".join_btn").classList.add("disable_btn"),document.querySelector(".join_btn").disabled=!0)}
