--- docs/assets/setDefaultData-b6PVXYnJ.js
@@ -1,1 +1,0 @@
-const a={isAuth:!1,user:null,token:""},t={user:null,id:[]};export{a,t as d};
