+++ docs/index.html
@@ -4,15 +4,14 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="/Karly/assets/favicon-pTJEpSkX.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script type="module" crossorigin src="/Karly/assets/main-9Oi0yUDK.js"></script>
+    <title>칼리 |</title>
+    <script type="module" crossorigin src="/Karly/assets/main-tChaEllP.js"></script>
     <link rel="modulepreload" crossorigin href="/Karly/assets/style-v06b81ne.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/pocketbase-mqgmocw_.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/header-MICfn_ID.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/getPbImageURL-Q86jnK4B.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/itemList-RegNhfZf.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-jFd8YWAz.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/header-tMjJ4wC4.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-PSdSBPgg.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/checkLogin-dd8Ug7jR.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/itemList-MN9rZj0T.js">
     <link rel="modulepreload" crossorigin href="/Karly/assets/setDocumentTitle-_OwlDZqT.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/setDefaultData-b6PVXYnJ.js">
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/main-nC5Dse61.css">
@@ -429,18 +428,12 @@
           <div
             class="swiper-button-prev banner"
             aria-label="메인 배너 슬라이더 이전으로"
+            role="button"
           ></div>
           <div
             class="swiper-button-next banner"
             aria-label="메인 배너 슬라이더 다음으로"
-          ></div>
-          <div
-            class="swiper-button-prev banner"
-            aria-label="메인 배너 슬라이더 이전으로"
-          ></div>
-          <div
-            class="swiper-button-next banner"
-            aria-label="메인 배너 슬라이더 다음으로"
+            role="button"
           ></div>
           <div class="swiper-pagination banner"></div>
         </div>
@@ -465,15 +458,18 @@
             <div
               class="swiper-button-prev recommend"
               aria-label="추천상품 슬라이더 이전으로"
+              role="button"
             ></div>
             <div
               class="swiper-button-next recommend"
               aria-label="추천상품 슬라이더 다음으로"
+              role="button"
             ></div>
             <div class="swiper-pagination recommend"></div>
           </div>
         </section>
         <section class="main_section line_banner">
+          <h2 class="sr_only">중간 배너</h2>
           <a href="">
             <figure>
               <img
@@ -494,10 +490,12 @@
             <div
               class="swiper-button-prev sale"
               aria-label="할인 상품 슬라이더 이전으로"
+              role="button"
             ></div>
             <div
               class="swiper-button-next sale"
               aria-label="할인 상품 슬라이더 다음으로"
+              role="button"
             ></div>
           </div>
         </section>
