+++ docs/src/components/header/index.html
@@ -12,7 +12,7 @@
     />
     <title>Document</title>
     <script type="module" crossorigin src="/Karly/assets/style-v06b81ne.js"></script>
-    <script type="module" crossorigin src="/Karly/assets/header-MICfn_ID.js"></script>
+    <script type="module" crossorigin src="/Karly/assets/header-tMjJ4wC4.js"></script>
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
   </head>
