+++ docs/src/pages/cart/index.html
@@ -11,15 +11,13 @@
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard.min.css"
     />
     <title>칼리 |</title>
-    <script type="module" crossorigin src="/Karly/assets/cart-CUSURSTP.js"></script>
+    <script type="module" crossorigin src="/Karly/assets/cart-69-rDkiH.js"></script>
     <link rel="modulepreload" crossorigin href="/Karly/assets/style-v06b81ne.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/header-MICfn_ID.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/header-tMjJ4wC4.js">
     <link rel="modulepreload" crossorigin href="/Karly/assets/footer-GuUMSODD.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/pocketbase-mqgmocw_.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/getPbImageURL-Q86jnK4B.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-jFd8YWAz.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-PSdSBPgg.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/checkLogin-dd8Ug7jR.js">
     <link rel="modulepreload" crossorigin href="/Karly/assets/setDocumentTitle-_OwlDZqT.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/setDefaultData-b6PVXYnJ.js">
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/footer-q-xIX-mF.css">
@@ -42,11 +40,11 @@
         <div class="header_inner">
           <ul class="header_member_service">
             <li>
-              <a href="/src/pages/register/">회원가입</a>
+              <a href="/Karly/src/pages/register/">회원가입</a>
             </li>
             <li class="divider" aria-hidden="true"></li>
             <li>
-              <a href="/src/pages/login/">로그인</a>
+              <a href="/Karly/src/pages/login/">로그인</a>
             </li>
             <li class="divider" aria-hidden="true"></li>
             <li>
@@ -96,7 +94,7 @@
                     </p>
                     <div class="location_modal_btn">
                       <a
-                        href="/src/pages/login/"
+                        href="/Karly//src/pages/login/"
                         class="location_modal_btn_login"
                         >로그인</a
                       >
@@ -110,7 +108,7 @@
                   </a>
                 </div>
                 <div>
-                  <a href="/src/pages/cart/">
+                  <a href="/Karly//src/pages/cart/">
                     <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACQAAAAkCAYAAADhAJiYAAAACXBIWXMAAAsTAAALEwEAmpwYAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAJLSURBVHgB7VbLbeJQFL2AWY+nA08F4+EjsbPpIFMBmQqSqSDpIEoFIRXEqSBmhwRITgfuIOz55Rx0nTgGHP/CIuJI5tnXj+vz7l/khBO+GWpJQa/Xs5bLpRWXGYYRjsfjUI4AIylYLBbXWAYJmY+lL0fAjoVs23br9br1tqFWu6B4tVr9DIJgLl+MHQvho378udPpmJvNxm40Gi4ePfli1D7bAEIuCD3BUgHWQKrF42w283IRItrt9gsWU6rHEIT+xQVGxj/SMoyjX1XEEQ54huUBFh8l39WzqZBnXCbS35Zq4PIHcRkUIoT4ifxcFSEH13wymRQjhEK5/SNM7EpJoKwwFm0myb73mQhp3FCBIyURc/ujFCWkYACabC1SDq6uxS1ERDUI7juTEoCerZWn06kvZQit12tPFf6WcmD8+IdeZibEOAKZUN5Nnhus+ljYip4P7claGLfAyRhHAyrmSCI5AXc7qserihD72YC9DSOJFEVURkoTgqIhqquF8eSHFATddYwxpjJk6vZJ6EgygAvnuO73tYAIrVbrHHsc7kXc3X42CucmpB+4S4j/Juca3XuDvZcxUdhsNvtppPJU6i3wgSsqRhz94TiC+7nKPqDb7dpKxuP4C4v+x72FZLhM05+bEJXS/HQTEOKZmbczBUBm6gFGDGIU1qG+Sk2IXFmm8Nn16Q59dvdVXqY2MpLF9AJ7ScxRoqM05bljCOODhQ894dZSUQiX9NVaH7An3u4Ra+dSJaEI2gYONskIegAbMRemZeMJJ3xbvALOGwU5v8fS6gAAAABJRU5ErkJggg==" alt="장바구니" />
                   </a>
                 </div>
@@ -218,10 +216,10 @@
                 </a>
               </div>
             </li>
-            <li><a href="/src/pages/product_list/">신상품</a></li>
-            <li><a href="/src/pages/product_list/">베스트</a></li>
-            <li><a href="/src/pages/product_list/">알뜰쇼핑</a></li>
-            <li><a href="/src/pages/product_list/">특가/혜택</a></li>
+            <li><a href="/Karly/src/pages/product_list/">신상품</a></li>
+            <li><a href="/Karly/src/pages/product_list/">베스트</a></li>
+            <li><a href="/Karly/src/pages/product_list/">알뜰쇼핑</a></li>
+            <li><a href="/Karly/src/pages/product_list/">특가/혜택</a></li>
             <li>
               <button class="menu_dawn">샛별·낮 <span>배송안내</span></button>
             </li>
@@ -325,10 +323,10 @@
                 </a>
               </div>
             </li>
-            <li><a href="/src/pages/product_list/">신상품</a></li>
-            <li><a href="/src/pages/product_list/">베스트</a></li>
-            <li><a href="/src/pages/product_list/">알뜰쇼핑</a></li>
-            <li><a href="/src/pages/product_list/">특가/혜택</a></li>
+            <li><a href="/Karly/src/pages/product_list/">신상품</a></li>
+            <li><a href="/Karly/src/pages/product_list/">베스트</a></li>
+            <li><a href="/Karly/src/pages/product_list/">알뜰쇼핑</a></li>
+            <li><a href="/Karly/src/pages/product_list/">특가/혜택</a></li>
             <li>
               <form action="/" class="search_form">
                 <label for="searchScroll"></label>
@@ -353,7 +351,7 @@
                     </p>
                     <div class="location_modal_btn">
                       <a
-                        href="/src/pages/login/"
+                        href="/Karly/src/pages/login/"
                         class="location_modal_btn_login"
                         >로그인</a
                       >
@@ -367,7 +365,7 @@
                   </a>
                 </div>
                 <div>
-                  <a href="/src/pages/cart/">
+                  <a href="/Karly/src/pages/cart/">
                     <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACQAAAAkCAYAAADhAJiYAAAACXBIWXMAAAsTAAALEwEAmpwYAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAJLSURBVHgB7VbLbeJQFL2AWY+nA08F4+EjsbPpIFMBmQqSqSDpIEoFIRXEqSBmhwRITgfuIOz55Rx0nTgGHP/CIuJI5tnXj+vz7l/khBO+GWpJQa/Xs5bLpRWXGYYRjsfjUI4AIylYLBbXWAYJmY+lL0fAjoVs23br9br1tqFWu6B4tVr9DIJgLl+MHQvho378udPpmJvNxm40Gi4ePfli1D7bAEIuCD3BUgHWQKrF42w283IRItrt9gsWU6rHEIT+xQVGxj/SMoyjX1XEEQ54huUBFh8l39WzqZBnXCbS35Zq4PIHcRkUIoT4ifxcFSEH13wymRQjhEK5/SNM7EpJoKwwFm0myb73mQhp3FCBIyURc/ujFCWkYACabC1SDq6uxS1ERDUI7juTEoCerZWn06kvZQit12tPFf6WcmD8+IdeZibEOAKZUN5Nnhus+ljYip4P7claGLfAyRhHAyrmSCI5AXc7qserihD72YC9DSOJFEVURkoTgqIhqquF8eSHFATddYwxpjJk6vZJ6EgygAvnuO73tYAIrVbrHHsc7kXc3X42CucmpB+4S4j/Juca3XuDvZcxUdhsNvtppPJU6i3wgSsqRhz94TiC+7nKPqDb7dpKxuP4C4v+x72FZLhM05+bEJXS/HQTEOKZmbczBUBm6gFGDGIU1qG+Sk2IXFmm8Nn16Q59dvdVXqY2MpLF9AJ7ScxRoqM05bljCOODhQ894dZSUQiX9NVaH7An3u4Ra+dSJaEI2gYONskIegAbMRemZeMJJ3xbvALOGwU5v8fS6gAAAABJRU5ErkJggg==" alt="장바구니" />
                   </a>
                 </div>
