+++ docs/src/pages/login/index.html
@@ -11,12 +11,11 @@
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard.min.css"
     />
     <title>칼리 |</title>
-    <script type="module" crossorigin src="/Karly/assets/login-52VG-BlH.js"></script>
+    <script type="module" crossorigin src="/Karly/assets/login-_cfNKZVO.js"></script>
     <link rel="modulepreload" crossorigin href="/Karly/assets/style-v06b81ne.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/header-MICfn_ID.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-jFd8YWAz.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/header-tMjJ4wC4.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-PSdSBPgg.js">
     <link rel="modulepreload" crossorigin href="/Karly/assets/setDocumentTitle-_OwlDZqT.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/pocketbase-mqgmocw_.js">
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/login-kHxLQ7FL.css">
