+++ docs/src/pages/product_list/index.html
@@ -11,12 +11,12 @@
     />
     <link rel="icon" href="/Karly/assets/favicon-pTJEpSkX.ico" />
     <title>칼리 | 베스트</title>
-    <script type="module" crossorigin src="/Karly/assets/product_list-guHp0qGC.js"></script>
+    <script type="module" crossorigin src="/Karly/assets/product_list-9G2DO7rr.js"></script>
     <link rel="modulepreload" crossorigin href="/Karly/assets/style-v06b81ne.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/header-MICfn_ID.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/getPbImageURL-Q86jnK4B.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/itemList-RegNhfZf.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/pocketbase-mqgmocw_.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/header-tMjJ4wC4.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/storage-PSdSBPgg.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/checkLogin-dd8Ug7jR.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/itemList-MN9rZj0T.js">
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/product_list-4j9DCiNE.css">
@@ -500,7 +500,7 @@
         </div>
         <ul class="item_list_container">
           <li class="item_card_container">
-            <a href="/Karly/src/pages/product_detail/">
+            <a href="/src/pages/product_detail/">
               <figure>
                 <img
                   class="product_img"
@@ -515,16 +515,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -579,16 +577,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -643,16 +639,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -707,16 +701,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -771,16 +763,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
