+++ docs/src/pages/register/index.html
@@ -15,11 +15,10 @@
       src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"
     ></script>
     <title>칼리 |</title>
-    <script type="module" crossorigin src="/Karly/assets/register-qnNIt04X.js"></script>
+    <script type="module" crossorigin src="/Karly/assets/register-61AnPVx2.js"></script>
     <link rel="modulepreload" crossorigin href="/Karly/assets/style-v06b81ne.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/header-MICfn_ID.js">
+    <link rel="modulepreload" crossorigin href="/Karly/assets/header-tMjJ4wC4.js">
     <link rel="modulepreload" crossorigin href="/Karly/assets/setDocumentTitle-_OwlDZqT.js">
-    <link rel="modulepreload" crossorigin href="/Karly/assets/pocketbase-mqgmocw_.js">
     <link rel="stylesheet" crossorigin href="/Karly/assets/style-2xt-1IBb.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/header-tovBoa8h.css">
     <link rel="stylesheet" crossorigin href="/Karly/assets/register-p5ZQtHJP.css">
