+++ src/lib/dom/index.js
@@ -8,4 +8,5 @@ export * from './itemList.js';
 export * from './showAlert.js';
 export * from './bindEvent.js';
 export * from './endScroll.js';
+export * from './checkLogin.js';
 export { default as clearContents } from './clear.js';
