+++ src/pages/cart/index.html
@@ -32,11 +32,11 @@
         <div class="header_inner">
           <ul class="header_member_service">
             <li>
-              <a href="/src/pages/register/">회원가입</a>
+              <a href="%BASE_URL%src/pages/register/">회원가입</a>
             </li>
             <li class="divider" aria-hidden="true"></li>
             <li>
-              <a href="/src/pages/login/">로그인</a>
+              <a href="%BASE_URL%src/pages/login/">로그인</a>
             </li>
             <li class="divider" aria-hidden="true"></li>
             <li>
@@ -86,7 +86,7 @@
                     </p>
                     <div class="location_modal_btn">
                       <a
-                        href="/src/pages/login/"
+                        href="%BASE_URL%/src/pages/login/"
                         class="location_modal_btn_login"
                         >로그인</a
                       >
@@ -100,7 +100,7 @@
                   </a>
                 </div>
                 <div>
-                  <a href="/src/pages/cart/">
+                  <a href="%BASE_URL%/src/pages/cart/">
                     <img src="/src/assets/icons/Cart.png" alt="장바구니" />
                   </a>
                 </div>
@@ -208,10 +208,10 @@
                 </a>
               </div>
             </li>
-            <li><a href="/src/pages/product_list/">신상품</a></li>
-            <li><a href="/src/pages/product_list/">베스트</a></li>
-            <li><a href="/src/pages/product_list/">알뜰쇼핑</a></li>
-            <li><a href="/src/pages/product_list/">특가/혜택</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">신상품</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">베스트</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">알뜰쇼핑</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">특가/혜택</a></li>
             <li>
               <button class="menu_dawn">샛별·낮 <span>배송안내</span></button>
             </li>
@@ -315,10 +315,10 @@
                 </a>
               </div>
             </li>
-            <li><a href="/src/pages/product_list/">신상품</a></li>
-            <li><a href="/src/pages/product_list/">베스트</a></li>
-            <li><a href="/src/pages/product_list/">알뜰쇼핑</a></li>
-            <li><a href="/src/pages/product_list/">특가/혜택</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">신상품</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">베스트</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">알뜰쇼핑</a></li>
+            <li><a href="%BASE_URL%src/pages/product_list/">특가/혜택</a></li>
             <li>
               <form action="/" class="search_form">
                 <label for="searchScroll"></label>
@@ -343,7 +343,7 @@
                     </p>
                     <div class="location_modal_btn">
                       <a
-                        href="/src/pages/login/"
+                        href="%BASE_URL%src/pages/login/"
                         class="location_modal_btn_login"
                         >로그인</a
                       >
@@ -357,7 +357,7 @@
                   </a>
                 </div>
                 <div>
-                  <a href="/src/pages/cart/">
+                  <a href="%BASE_URL%src/pages/cart/">
                     <img src="/src/assets/icons/Cart.png" alt="장바구니" />
                   </a>
                 </div>
