+++ src/pages/product_list/index.html
@@ -495,7 +495,7 @@
         </div>
         <ul class="item_list_container">
           <li class="item_card_container">
-            <a href="%BASE_URL%src/pages/product_detail/">
+            <a href="/src/pages/product_detail/">
               <figure>
                 <img
                   class="product_img"
@@ -510,16 +510,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -574,16 +572,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -638,16 +634,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -702,16 +696,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
@@ -766,16 +758,14 @@
                 </figcaption>
               </figure>
             </a>
-            <button class="cart_button">
+            <button class="cart_button" aria-label="담기">
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 width="45"
                 height="45"
                 viewBox="0 0 45 45"
                 fill="none"
-                aria-labelledby="cart"
               >
-                <title id="cart">담기</title>
                 <path
                   opacity="0.5"
                   d="M22.5 45C34.9264 45 45 34.9264 45 22.5C45 10.0736 34.9264 0 22.5 0C10.0736 0 0 10.0736 0 22.5C0 34.9264 10.0736 45 22.5 45Z"
