+++ src/pages/product_list/product_list.js
@@ -1,8 +1,17 @@
 import '/src/components/header/header.js';
 import '/src/components/footer/footer.js';
 import '/src/pages/product_list/product_list.css';
-import { getNode, getNodes, insertLast, css, renderItemList } from '/src/lib';
+import {
+  getNode,
+  getNodes,
+  insertLast,
+  css,
+  renderItemList,
+  checkLogin,
+} from '/src/lib';
 import pb from '/src/api/pocketbase';
+
+checkLogin();
 
 // 포켓호스트 서버에서 상품 정보 받아와서 제품 목gi록 페이지 화면에 렌더링
 const itemList = await pb.collection('products').getList(1, 15);
